[bt][host] Handle IOCTLs being called whilst device is unbinding
It is possible to observe a ZX_DEBUG_ASSERT fail when rapidly creating
and destroying bt-fake-hci devices; this appears to be due to bt-gap
calling the OpenHostChannel method whilst the device is in the process of
unbinding, revealing a race condition where `host_` has been set to null
but the OpenHostChannel call still attempts to use it.

This change causes the driver to check for that condition, and return
ZX_ERR_BAD_STATE in that case.

Change-Id: I8de51d1319b19a61372b6087bf9a2233f832e106