[bt][bredr] Fix Inquiry and Remote Name Conflict
Inquiry requires the link manager to be in the inquiry scan substate,
and Remote Name Request (when not connected) will create a temporary
link layer connection to request the remote name.  This temporary
connection requires the link manager to be in the page scan substate,
which is disallowed while it is in the inquiry scan substate.

Exclude these commands from being run simultaneously using the
CommandChannel support to prevent commands running at the same time.

BT-284 #done
Test: use bt-cli to start discovery, inspect bt-snoop and logs

Change-Id: I90746deb6784787f86e9a002fc205419c30bf436