[bt][l2cap] Make failed channels into invalid sockets
When building a socket for a channel with the socket version of
OpenL2capChannel, make SocketFactory return an invalid socket when
passed a channel that had failed to open.

The current behavior is to assert and fail on the null channel, which
does not match the contract of the channel version of OpenL2capChannel.

BT-942 #done

Test: in bt-host-unittests,
DATA_SocketFactoryTest.SocketCreationFailsIfChannelIsNullptr

Change-Id: I231eaa7e3fe0c614fc432dec4d33312158e43dc0