[bt] Support write long char/desc values
BT-318 #done

Testing:
bt-host-unittests : Added unit tests for the gatt client methods and
added unit tests for the higher level service logic as well.
Tested an actual long write to a device - device seemed to accept it.
Tested with long writes in PTS! Verified that we sent prepare writes and
followed with execute write as expected.

Change-Id: I4b66efc0ef862ea0d671298bc0dcd5cbe14a9ce7