[bt][host] Add Host.Disconnect
Add the Disconnect method to the fuchsia.bluetooth.host fidl library,
and implement it in the bt-host driver.

Currently, a disconnect attempt made whilst connecting to the named
device will fail. In the future (bt-873), this will result in the
connection attempt being cancelled.

The host driver integration test harness has been reworked to use the
generic Harness type which is used for other harnesses. As the
HostDriver harness pre-dated the generic harness, it had not previously
been updated to the new standard.

Test: Integration tests have been added to exercise the disconnect
functionality, they can be run with the usual command for integration
tests:

  fx run-test bluetooth-tests -t bt-integration-tests

BT-652 #done

Change-Id: I4cdf605af98b3a5cdf6e2c7d3ab658824410159c