[bt][sdp] Limit ServiceSearch response length
While not used often, PTS requires proof that ServiceSearch response can
generate a continuation state.  It does this by setting the MTU size to
48.  We respond with succes from this but do not support actually
limiting MTU (BT-616).

We are allowed to shorten any response, so set the maximum amount
of services returned to 9 to be below the MTU set.

Add support for generating ServiceSearch responses with continuation
state.  Default is 9 services to produce a maximum response of 48 bytes.

Modified PDU tests to be less noisy when passing.

BT-644 #done
BT-2266 #done

Test: - added SDP_ServerTest.ServiceSearchContinuationState
  - modified SDP_PDUTest.ServiceSearchResponsePDU
Change-Id: I62ca96e217a2451e3c089e0f869e44d83a20d4ab