[bt][gap-bredr] Add callback to PairingState::InitiatePairing
Add a callback that allows local pairing initiators to wait for pairing
status, e.g. in order to proceed with Security Mode 4 outbound L2CAP
service access.

Test that the callback is signaled even if we initiated pairing while
the peer had already began to pair with us (i.e. we are responder).

BT-374

Test: in bt-host-unittests, GAP_PairingStateTest.*

Change-Id: Ibe5e1edcffbcec82b945eb02d501c923d3068171