[bt][hci] Add RemoveQueuedCommand to CommandChannel
Add function for the caller of SendCommand to try to remove the command
from the send queue for the Bluetooth controller. This allows the user
of CommandChannel to cancel a queued command or—if it is already sent—
to cancel it (as necessary) in an HCI-specific way.

Add DEBUG log output for sending queued commands.

Extract MakeReadRemoteSupportedFeatures test code.

Bug: 668

Test: bt-host-unittests HCI_CommandChannelTest.RemoveQueued*Command*

Change-Id: I9ccbc78c6649276ee544861bce603b6171b152aa