[bt][sm] Add SecurityProperties::GetLinkKeyType
Add a member function to SecurityProperties that converts it to the
hci::LinkKeyType that would produce those properties.

This will be used to obtain the link key type from stored bonds.

This, combined with the ctor that takes hci::LinkKeyType, allows the GAP
Connection state bookkeeping to avoid using SMP types in its code, in
order to maintain layering.

BUG: 967

Test: in bt-host-unittests,
SMP_TypesTest.SecurityPropertiesToLinkKeyType

Change-Id: Ieaf6f8bc83f03021110e1d958092f03a5ac2addf