[bt][gap] Handle SDP channel creation failure
Previously when an L2CAP channel was created for SDP,
the callback failed to check if the channel creation failed
(is null), and the null channel was passed to the SDP client.
The SDP client tries to dereference the channel, which results in
a crash.

This fix skips SDP client creation and service discovery if channel
creation fails.

Bug: 36508

Test: in bt-host-unittests
GAP_BrEdrConnectionManagerTest.SDPChannelCreationFailsGracefully

Change-Id: I9dd2c1e8d208c8e7efb177fe179017ca269ee8fd