[bt][gap] Specify role in PairingDelegate::DisplayPasskey
With Numeric Comparison association, both sides display a passkey and
accepts the user's confirmation to pair.

Add an argument to DisplayPasskey for "display method," which selects
PairingMethod::PASSKEY_COMPARISON or PASSKEY_DISPLAY to ask for the
local user to confirm pairing or enter the passkey on the peer,
respectively.

Extract calls to PairingDelegate interface into a DisplayPairingRequest
method.

Bug: 967

Test: in bt-host-unittests,
HostServerTest.HostDisplayPasskeyRequestsPasskeyDisplayOrNumericComparisonPairingOverFidl

Change-Id: Iabff90c8c076e8ba3c0a69aa715b7c5598c79aa8