[bt] Add type restrictions to ByteBuffer casts
In ByteBuffer, reinterpret_cast is used to perform conversions between
generic types and byte representations. This adds restrictions that the
generic type must be TriviallyCopyable, which roughly maps to the
ability to relocate the byte representation of an object safely.

Additionally, because MutableByteBuffer::WriteObj can deduce its generic
type, disallow that type from being some kind of ByteBuffer because the
user likely does not want to write the byte representation of ByteBuffer
itself.

Bug: 41139
Test: Build and run bt-host-unittests. Manually test that
non-TriviallyCopyable types cause compilation errors for As and
WriteObj. Test that WriteObj<BufferView> gives an error message about
how to fix it.

Change-Id: Ib5a7fa1a3ccf648a070565931e6c704e118919bb