Minor pw_bloat updates
This change renames the bloat_report template to pw_size_report as it is
more explicit.

Additionally, the option to override the global Bloaty config file on a
per-report basis is removed. This is due to Pigweed's policy of having
one target per build invocation. The Bloaty config is inherently tied to
an executable build target and linker script, both of which are provided
by the target. Overriding the Bloaty config without changing the target
can result in meaningless report output.

Change-Id: I1fa444d065d964a3a85ade50039e06ae807050f2