[bt][sdp] Respect the TX MTU of the channel
SDP packets could be generated that were larger than the MTU allowed by
the channel before.  Pass the maximum PDU size through to the GetPDU
methods which are encoding the PDU for the channel, and automatically
generate continuation state when the response will not fit.

Update Channel to take an argument of the MTU which is set at startup.

Fixed: 38500
Bug: 1022

Test: added SDP_PDUTest.ServiceSearchAttributeResponseGetPDU_MaxSize,
      added SDP_PDUTest.ServiceAttributeRespondeGetPDU_MaxSize,
      added SDP_PDUTest.ServiceSearchRespondeGetPDU_MaxSize,

Change-Id: I8407ce0ac6a30990fd373718265f2923770b9aef