pw_cmd: Add automatic plugin loading
This adds automatic detection and loading of 'pw' CLI plugins. For now,
the mechanism is to look for modules named 'pw_*' with a
'PW_CLI_PLUGINS' variable containing a list of Command classes.

In the future we may want to explore explicit plugin registration, which
is faster than scanning over all modules like this version does.

Change-Id: I29073e556189abc8d408fdf23257554bbda45ba5