Check Python version after env_setup/cipd/init.sh.
Check Python actually comes from CIPD in env_setup/setup.sh. This
produces an error if something failed with CIPD and the CIPD init
script failed to detect it.

Check Python version is 3.8 in env_setup/virtualenv/init.py. If not
Python 3.8, print an error saying the version is incorrect.

Bug: 23
Change-Id: I7d92cb9838d2c6d6941038a229c875a6f7f1d4ff