[bt][gap][fidl] Add non-bondable mode plumbing to bt-host/{fidl,gap}
Enables the GAP bondable mode, implemented in bt-host/sm, to be
controlled from the higher layers of bt-host i.e. FIDL servers.
Functional changes made to LowEnergyConnectionManager consist of passing
the bondable mode setting from the public functions of the
LowEnergyConnectionManager to the SecurityManager of one of its
connections through a series of callbacks.

Bug: 35584
Test: Run them with `$ fx run-test bluetooth-tests -t bt-host-unittests\
      -- --gtest_filter="GAP_LowEnergyConnectionManager*"`. Tests added
      to GAP_LowEnergyConnectionManagerTest::{
      	 PassBondableModeThroughRemoteInitiatedLink,
	 PassNonBondableThroughRemoteInitiatedLink,
	 PassBondableThroughConnect,
	 PassNonBondableThroughConnect,
      }
Change-Id: I090e879bf10f8892e58c9ce6baac58b55bc11cef