[bt][l2cap] Add channel parameters to Channel Manager
Add channel parameters to outbound and inbound channel creation
interfaces of ChannelManager and LogicalLink.

Create l2cap/types.h file for non-spec type declarations used throughout
the stack.

Refactored channel callbacks in ChannelManager and
DynamicChannelRegistry into ServiceInfo struct. ServiceInfo is
parameterized on the type of channel callback, as ChannelManager and
DynamicChannelRegistry pass different types of channels to
the layers above them. ServiceInfo also stores the channel parameters
that are received when a service is registered, so that the parameters
can be used in channel configuration for that service.

Bug: 872

Test: in bt-host-unittests
L2CAP_ChannelManagerTest.OutboundChannelConfigurationUsesChannelParameters
L2CAP_ChannelManagerTest.InboundChannelConfigurationUsesChannelParameters

Change-Id: I9b0342e3f53d2d9d1e53242159f3b01326536d49