Bazel: compile C files separately from C++
Some Pigweed compilation options do not work for C files, and Bazel
doesn't provide a way to split C and C++ compmile options at the
cc_library level. This change updates the pw build rules to build C
code, if any, in a separate cc_library that the C++ library depends on.

Change-Id: I7c1b3c95665a382e35c69f3ec57d4a1cca8abe2d