pw_protobuf: Start decoder class
This change adds a Decoder class to the pw_protobuf module for decoding
protobuf messages from their wire format. The decoder interface is event
based, with a user-defined handler responsible for processing each field
in the message as it is encountered.

Currently, only varint fields are supported in the decoder. The other
protobuf wire types will be added in a future change.

Change-Id: I5a16d1864bb082d0ac1ec759064c33bb27db3e72