pw_span: Fix deduction guides for containers
Previously, the constness of the container implied the constness of the
elements, which is not always correct. A mutable std::string has mutable
elements, but a mutable std::string_view has const elements. The result
was that template argument deduction failed in some cases, including a
mutable std::string_view.

This change has the container deduction guides use the actual element
type from the container as the element type for the span.

Change-Id: Idafde58b8cfedf86e1f90a3e22fb05780903222e