pw_presubmit: Add message about the Bazel cache
Sometimes the Bazel cache gets in a bad state where it complains about
undeclared inclusions on C standard library headers. The only way to fix
this is to delete the Bazel cache. This adds a message to the presubmit
about how to clear the cache if the Bazel build fails.

Fixed: 114
Change-Id: I7c3ad45bdce919a325f721f2071e99c1e56582c0