pw_watch: Add proper debouncing; enter to build
This adds a more sophisticated debounce mechanism to the watcher, which
restarts with each incoming file change. This prevents starting builds
too early when doing rebases or pulls which trigger many file changes,
and can cause Ninja or GN to get into an inconsistent state.

If filesystem changes happen during a build, the build is marked as
"interrupted" and failed, to prevent confusing inconsistent builds.

Additionally, this change runs a build when pressing 'enter'. This is
useful in many cases where something outside of the watcher's view
changes that should trigger a rebuild; for example, when first running
pw watch.

Change-Id: Ifa6e36eeaf742f93ce451115f18691490e6085d4