[bt][sm] Implement ValidPacketReader & PairingPhase
PairingPhase is the base class from which all other Pairing Phases are
derived. A "PairingPhase" is one of the phases described in Vol. 3 Part
H Section 2.1, or the phase in which no pairing is active. This base
PairingPhase class implements functionality relevant to all phases of
pairing. Concrete PairingPhases may use the ValidPacketReader class to
parse inbound SDUs into known-good SMP packets.

Bug: 46278
Test: Added 3 tests to pairing_phase_unittest.cc, as well as fake_phase_
      listener.h class for use with unit testing the rest of the pairing
      phases. Also added packet_unittest with 5 packet parsing tests.
      Run the new tests with `PacketTest` or `PairingPhaseTest` in place
      of [suite_name] in the below command:
      `fx run-test bluetooth-tests -t bt-host-unittests -- \
       --gtest_filter="SMP_[suite_name]*" `
Change-Id: I5c2d805bb960d389c454b4cbfcfdde6f39b74473