pw_kvs: Don't garbage collect sectors with copies of an in-progress key
Don't garbage collect sectors in the addresses to avoid list. These are
addresses of redundant copies already written of the key we are trying
to write.

This results is failing to write keys sooner in some workloads with
near-full flash.  This change is to simplify how GC works to make some KVS restructuring
easier to do.

Change-Id: I1bc52728fa29006253dbff529447e7385376d3a7