[bt][l2cap] Support sending Connection Parameter Update Request
Add LogicalLink::SendConnectionParameterUpdateRequest, which sends a
Connection Parameter Update Request on the LE signaling channel. The
result in the response is converted into a bool and passed to a result
callback. A Command Reject Response is treated as a rejection result.

Add ChannelManager::RequestConnectionParameterUpdate, which calls
LogicalLink::SendConnectionParameterUpdateRequest on the link of a
specified connection handle.

Bug: 46968
Test: in bt-host-unittests
L2CAP_ChannelManagerTest.RequestConnParamUpdateForUnknownLinkIsNoOp
L2CAP_ChannelManagerTest.RequestConnParamUpdateAsSlaveAndReceiveAcceptedAndRejectedResponses
L2CAP_ChannelManagerTest.ConnParamUpdateRequestRejected

Change-Id: I5d7fa1e1c172f1081b8956821bcaacac8e79e918