pw_env_setup: fix PATH cleanup
Previously, values were removed from PATH on Windows by searching for
";value" or "value;" and replacing with the empty string. This worked
for values that weren't prefixes of other values. If PATH was set to
"C:\\abc;C:\\foo;C:\\foo\\bar;C:\\xyz" then removing "C:\\foo" should
result in "C:\\abc;C:\\foo\\bar;C:\\xyz" but it actually resulted in
"C:\\abc\\bar;C:\\xyz".

This commit removes the ";value" search so we only replace "value;"
with the empty string. And since all our values start with "C:\\" we
always start at the beginning of the value. A downside is we'll never
remove values at the end of PATH, but I think that's ok.

Change-Id: Ia138213fcb452ccf3219de915013fc9c7c7a4638
Fixed: 162