Allow-list unsigned integer overflows.
This CL adds the 'no_sanitize("unsigned-integer-overflow")' Clang/LLVM
attribute to several locations that intentionally rely on that behavior.

From https://clang.llvm.org/docs/UndefinedBehaviorSanitizer.html: "this
is not undefined behavior, but it is often unintentional." By using an
explicit allow-list, we should still be able to detect the latter case.

Change-Id: I7d7da40c73fd22b6b29196dc4788eb00fca61439
Bug: 190