pw_containers: Don't downcast vectors to larger sizes
This CL removes two checks in vector_test that explciitly downcast a
Vector<int> to a Vector of a larger size. It also modifies the data()
accessors to downcast to a size of 0 instead of 1, as that's the only
size guaranteed to be less than or equal to the true vectors size.

Downcasting to a larger size is undefined behavior, as detected by
Clang/LLVM's -fsanitize=object-size

Change-Id: Idd87e968d10174ed49edf470833519601cdbbe35
Bug: 188