[bt] Create data domain fuzzer
Create an overall fuzzer of the data, hci, and l2cap stack.

Fuzzer data is received as ACL data packets and also triggers the
opening and closing of a connection (so that some packets are valid).
One service is registered to enable the possibility of the fuzzer
opening an l2cap channel.

Don't fuzz the length field of ACL data header because it is set by the
controller and we are confident we handle it correctly. Most fuzzed
length fields would be incorrect, and so most packets would be rejected
immediately, which would slow down fuzzing.

Don't fuzz the handle field of ACL data header in order to speed up
fuzzing. Use handle of connection that is toggled by fuzzer instead.

Bug: 51794
Test: (only adds test code)

Change-Id: I04f316302ca449e5eca734fd508d174e8798ac14
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387759
Testability-Review: Ben Lawson <benlawson@google.com>
Commit-Queue: Ben Lawson <benlawson@google.com>
Reviewed-by: Matthew Dempsky <mdempsky@google.com>