[bt][l2cap] Fix ERTMRxEngine tests, clean up docs
Fix bug in ProcessPduCallsRemoteBusySetCallbackOnReceiverNotReady and
ProcessPduCallsRemoteBusyClearedCallbackOnReceiverReadyAfterReceiverNotReady
that checked for the ordering between RemoteBusy callbacks and the
AckSeqNum callback, which is not relevant. The latter callback is for a
local variable when it should be checking against the ReceiveSeqNum
callback which is for the peer variable.

Additionally, add docs and rename some tests.

Bug: 1353
Test: in bt-host-unittests, update
L2CAP_EnhancedRetransmissionModeRxEngineTest.*

Change-Id: I6b6b152ddc48d953e3b1ee209f204946e8edf51b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389963
Commit-Queue: Xo Wang <xow@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>
Testability-Review: Ben Lawson <benlawson@google.com>