[bt][sm] Minor patches to already-landed phase classes
V5.0 Vol. 3 Part H 3.6.1 says to ignore the EncKey bit for Secure
Connections pairing. This could be done by explicitly checking for
Secure Connections and ignoring the EncKey bit in the KeyDistGen fields
whenever they're accessed, but it is simpler to ignore the EncKey bit
by setting it to 0 when resolving the negotiated features in Phase 1.

Also adds an assertion to Phase 3 that it has not failed when its public
methods are called.

Bug: 46278
Test: `fx test bt-host-unittests -- --gtest_filter="SMP*"`
      Added test SMP_Phase1Test.FeatureExchangeScIgnoresEncKeyBit.
Change-Id: I308ba6ceee87e9b3996cb01f23242d6a6469174f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386215
Commit-Queue: Lucas Jenkins <lucasjenkins@google.com>
Reviewed-by: Nick Pollard <nickpollard@google.com>
Testability-Review: Nick Pollard <nickpollard@google.com>