[bt][host] Fix inquiry scan interval
On both Intel and Broadcom controllers the bt-host driver always logs the
following warning:

[gap-bredr:bredr_discovery_manager.cc:503] write inquiry scan activity failed: [status: invalid HCI command parameters (HCI 0x12)]

This happens because the controller rejects the parameters given to the Write Inquiry Scan Activity
command during bt-host initialization. This is due to a mathematical error, in which the intended
300ms interval maps to 0x1E0 in jiffies but was incorrectly coded as 0x1E1 which does not map to a
whole number in milliseconds.

Test: 1. bt-host-unittests.
      2. bt-host should no longer log the message mentioned above.
      3. Discoverable mode should continue to work as expected.
Change-Id: I8b2d146b4e957dec7e32f3105ace4ae6efa53d70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396070
Reviewed-by: Ben Lawson <benlawson@google.com>
Testability-Review: Ben Lawson <benlawson@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>