[bt] Remove ununsed RFCOMM code
As part of the plan to move RFCOMM support to the other side of the
Profile API, this code is being removed.

It was not in use anyuhere, and bugs in the code are not going to be
addressed before the transition.

Bug: 49073
Fixed: 50629
Fixed: 37806

Test: only removed code, bt-host-unittests
Change-Id: Ic75672386f65053b94eb80c716c277068faf4cb8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398534
Testability-Review: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>