pw_kvs: Set KVS test to use any flash partition
Break out EmptyInitializedKvs to a seperate file and setup to use a
flash partition that is give by a getter method. The getter is defined
seperatly to allow reuse of the test file in other tests.

Add several test flash partitions and setup KVS and flash tests to use
them.

Change-Id: I8e28d71b3528b471c79518e658cb30c33d2b30c6
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/13620
Commit-Queue: David Rogers <davidrogers@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>