[bt][fidl] Create helper function to convert ServiceDefinition to ServiceRecord
Move code that converts a ServiceDefinition to a ServiceRecord from
profile_server.Advertise() to separate functions in the FIDL helper
library so that they can be accessed outside the profile server.

This is useful to allow for the creation and use of a SDP search results
field in fuchsia.bluetooth.test.BredrPeerParameters.

Test: fx test bt-host-unittests -- --gtest_filter="SDP_ServerTest",
FidlHelperTest.ServiceDefinitionToServiceRecord
Bug: 57370

Change-Id: Ic555ca5199d781c1494b165303564586c5ee1804
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412062
Commit-Queue: Anusha Datar <anushadatar@google.com>
Testability-Review: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Xo Wang <xow@google.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>