[bt][sm] Make SM dependency-injectable
Introduce an abstract class with the production public interface of
sm::SecurityManager and only basic getters/setters a la l2cap::Channel,
moving the production implementation to a separate SecurityManagerImpl.
This is in preparation for a follow-on change which introduces an SM
test double for use in GAP unit tests.

Bug: 886
Test: Pure refactor, run `$ fx test bt-host-unittests` to verify nothing
      broke
Change-Id: I9a706268012eeba1c759842fdb642ada449019c8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414455
Commit-Queue: Lucas Jenkins <lucasjenkins@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>
Testability-Review: Ben Lawson <benlawson@google.com>