[bt][fidl] Implement AudioDirectionExt server
Implement AudioDirectionExt server that sends ACL priority command when
SetPriority() is called. This is a short-term hack, and the command will
be properly managed in l2cap in the future.

It is currently possible for profiles to interfere with each other if
they request conflicting priorities. This will be fixed with a more
robust conflict checking system in the future, but it shouldn't be an
issue for now as we only expect to create 1 A2DP channel with a peer at
a time.

Reverse the ACL priority command when the AudioDirectionExt channel is
closed.

Create hci::BcmAclPriorityCommandParams as a temporary definition of the
ACL priority command until proper DDK support is implemented.

Add handler for vendor commands to FakeController. Tests can register a
callback to be called with any vendor commands received.

Bug: 57163
Bug: 57727
Test: in bt-host-unittests
FIDL_ProfileServerTest_ConnectedPeer/PriorityTest.ConnectAndSetPriority/*
FIDL_ProfileServerTest_ConnectedPeer.InboundConnectAndSetPriority

Change-Id: I1fcbf3a0264f28612c787cab2b1bd41d85550a04
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414006
Commit-Queue: Ben Lawson <benlawson@google.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
Testability-Review: Marie Janssen 💖 <jamuraa@google.com>