Allow mac to use -Wl,--start-group
We only need to avoid this for Mac targets, not all builds on Mac hosts,
because Mac's linker doesn't recognize this option. Let the toolchain
definition deal with whether to enable it.

This fixes a source of build fails that only happen on Mac.

Change-Id: I6cba23e14c92ea27675564bcb3f6004c96f384ab
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/15980
Reviewed-by: Armando Montanez <amontanez@google.com>
Reviewed-by: Keir Mierle <keir@google.com>
Commit-Queue: Michael Spang <spang@google.com>