[bt][acl] Ignore out of bounds completed packets
When processing the number of completed packets, a larger number of
packets than was actually included in the packet could be specified,
causing data that was out of bounds to be accessed.

Calculate the number of packets' worth of data and do not process for
counts that are not provided.

Fixed: 58592
Test: modified ACLDataChannelTest in bt-host-unittests to exercise

Multiply: bt-host-unitests
Change-Id: I72760c2af8a48dd3e8a3f791b43e2d0186e87583
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418774
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>
Testability-Review: Ben Lawson <benlawson@google.com>