pw_web_ui: Improve error handling of WebSerialTransport
Don't throw an error if a user tries to reconnect to a port. Pass errors
in an 'errors' observable that consumers can subscribe to. Don't put the
'chunks' observable into an errored state, since that prevents it from
ever sending anything else again, and we often want to continue reading
despite an error.

Change-Id: I114240e1c3c0556fd88b7d4ec1b40fec34a8ccb3
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/17167
Reviewed-by: Max Koopman <koopman@google.com>
Reviewed-by: Paul Mathieu <paulmathieu@google.com>
Reviewed-by: Matthew Soulanille <msoulanille@google.com>
Commit-Queue: Matthew Soulanille <msoulanille@google.com>