[fit] Move thread_checker functionality from fxl to fit
thread_checker is a type that can introduce runtime checks that a
sequence of calls are made from the same thread. This is useful for
adding runtime safety checks to thread-hostile types. This moves an
implementation of this logic from the internal fxl library to libfit so
that it can be exposed in the SDK and used by other C++ code published
in the SDK. In particular, some of the thread-hostile types in the fidl
library such as fidl::InterfacePtr could use this type in the future.

This change also renames the type and member functions of this type to
match the fit library style and updates it to depend on std::thread
instead of pthreads.

Test: Unit tests moved from fxl unit tests to fit unit tests

Change-Id: I704ec59acc1162b591ca5d71dc3c2b74f56a704e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446935
Commit-Queue: James Robinson <jamesr@google.com>
API-Review: Adam Barth <abarth@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Testability-Review: George Kulakowski <kulakowski@google.com>
Testability-Review: Corey Tabaka <eieio@google.com>
Testability-Review: James Robinson <jamesr@google.com>