pw_build: Requre headers in facades
The purpose of a facade is to prevent dependency cycles caused by a
backend needing to depend on the headers of a facade target. If a facade
target doesn't expose headers, a facade is essentially useless. In those
cases, developers should just directly depend on the variable that
'backend' would have otherwise been set to.

Change-Id: I23bcf1b93cc3c02013339813aa3f4196a432370d
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/24540
Commit-Queue: Armando Montanez <amontanez@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>