pw_unit_test: Add library targets
This updates the pw_test and pw_test_group templates to create a .lib
target which contains the test code as a source_set that can later be
linked into a pw_test. pw_test_groups additionally have a new .bundle
target which builds an executable with all of the group's tests.

Change-Id: Ic77d9d3a02cd29487a9f82001d9bf3314da4fa27
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/24621
Reviewed-by: Ewout van Bekkum <ewout@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>