pw_rpc: Return FailedPrecondition when writing to closed ServerWriter
If a ServerWriter is closed, its AcquirePayloadBuffer function returns
an empty span. The nanopb ServerWriter would then try to encode into
this span and fail, returning a cryptic INTERNAL status. This updates
the nanopb writer to correctly return FAILED_PRECONDITION if Write is
called when it is closed.

Change-Id: I30c70f0ee2a7f7c6f44bb8aded9aafa2045b99a4
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/25260
Commit-Queue: Alexei Frolov <frolv@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>