pw_kvs: Fix static analysis failure in unittest
Resolve static analysis complaint of a possible nullptr dereference. The
failure mode here could not happen as the existing code was taking a
reference of a reference, but the comparison to nullptr caused the
analyzer to complain. Explicitly check to see that a comparison to a
valid pointer is made before continuing.

Change-Id: I0e66624b46ba0de9ef48e80f63e8c96d4aa24253
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/31682
Reviewed-by: Keir Mierle <keir@google.com>
Commit-Queue: Prashanth Swaminathan <prashanthsw@google.com>