[bt][gap] Always call pairing status callbacks
When connecting to a peer, it's possible for the pairing to complete
successfully, and then the peer disappears.  This manifested itself as
never getting called back BrEdrConnectionManager::OpenL2capConnection,
stalling connections from Profile clients.

Update the BrEdrConnectionManager::OpenL2capConnection API to accurately
reflect that the result always happens through the callback. Update
PairingState to finish incomplete local pairing requests when it is
destroyed, since it holds the callbacks while pairing is in progress.

Update the docs for PairingState to accurately reflect that the
status_cb on creation is not called on destruction, and the remaining
pending callbacks are.

Fixed: 1414
Test: BrEdrConnectionManagerTest.OpenL2capPairingFinishesButDisconnects
Test: PairingStateTest.UnresolvedPairingCallbackIsCalledOnDestruction
Test: fx test bt-host-gap-tests
Change-Id: Iade00489065a726802daf483141b250926681187
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484860
Reviewed-by: Ben Lawson <benlawson@google.com>
Reviewed-by: Xo Wang <xow@google.com>
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Fuchsia-Auto-Submit: Marie Janssen 💖 <jamuraa@google.com>