[bt] Add PeerCache fuzz test
Add lightweight fuzz harness that mutates a single peer in a PeerCache
to look for anomalous configurations of Peer state. This is intended to
rapidly exercise the Peer machinery and its fuzzer, which will be reused
by a fuzzer for fuchsia.bluetooth.sys.Access/WatchPeers.

Bug: fxbug.dev/66355, fxbug.dev/59064
Bug: b/166657466, b/173822255
Test: fx fuzz start bt-host-fuzzers/peer_cache_fuzzer
Change-Id: Ifdf3ac521023e2029b3b869bade74c24110dfcc3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/488320
Commit-Queue: Xo Wang <xow@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>