pw_router: Remove logging
This removes all log statements from the static router. This is done as
there is no guarantee for what pw_log backend is used. In some systems,
logs may flow directly through the router, resulting in recursive calls
that cause a crash.

The user of the router can choose to check and log its return values if
the operation is safe.

Change-Id: I3a94d9245fde8fd5d27a23ddd1713aba1e880bd8
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/33621
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Alexei Frolov <frolv@google.com>