pw_assert_basic: Fix build and minor bugfix
This commit fixes the bazel build files. This fix also addresses a bug
where an interaction between @io_bazel//src/main/tools:process-wrapper
and pw_assert_basic caused the output from `bazel test
//some_test_target` to be completely empty on assertion. This occured
because pw_assert_basic was 'aborting' causing the core to dump without
calling the destructors. Under normal circumstances this is fine as
stdout is synchronised. However as bazel's process wrapper binds
stdout, it results in stdout buffering, meaning that when
pw_assert_basic 'aborts' the buffered stdout is lost. The replacement
of 'abort' to 'exit' allows destructors for the stdout handles to be
called thus flushing the buffer out to bazels process-wrapper.
This is not a critical bug and is more of a nuisance as the message that
prints the line number that caused the assertion is lost when the
program aborts.

Change-Id: Ia1599db78c5caa878f3d7496e2b3009e3c44dcea
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/34621
Reviewed-by: Keir Mierle <keir@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Keir Mierle <keir@google.com>