pw_unit_test: Add test suite filter
This updates the unit test framework to accept an optional list of test
suites to run. If set, RUN_ALL_TESTS will only run test cases which
match the provided suites. Additionally, tracking of skipped and
disabled tests in a test run is added.

The unit test RPC service is also updated to accept a list of suites in
a test run request.

Change-Id: Ia1aefdcf2314c24431bff75378c30fcdfdd24c27
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/40061
Commit-Queue: Alexei Frolov <frolv@google.com>
Reviewed-by: Keir Mierle <keir@google.com>
Reviewed-by: Joe Ethier <jethier@google.com>