pw_stream: Allow staging data to MemoryWriter end
Changes the memory writer to use std::memmove to allow data to be staged
to the unused portion of a MemoryWriter's buffer before it is written.
This is useful for operations that require direct access to a buffer to
build data before it is flushed to a MemoryWriter.

No-Docs-Update-Reason: Optimization that shouldn't be advertised
Change-Id: I3728e50e30031d7f87db43981827ef799fbce217
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/40244
Reviewed-by: David Rogers <davidrogers@google.com>
Commit-Queue: Armando Montanez <amontanez@google.com>