[bt][gap-bredr] Log early initiator pairing failures
In PairingState, log any error code that is in an Authentication
Complete event, instead of emitting only a generic "unexpected event"
error. This case occurs when a peer is not ready to pair or timed out
during user-interactive pairing.

Additionally, should PairingState cause a link to be disconnected, log
the Authentication Complete status code in the "link not found" case as
well.

Document this possibility in the header comments.

Bug: fxbug.dev/74237
Test: fx test //src/connectivity/bluetooth/core/bt-host
In bt-host-gap-tests,
GAP_PairingStateTest.AuthenticationCompleteWithErrorCodeReceivedEarlyFailsPairing
covers the new behavior of reporting the early exit error code.
Otherwise the change is logging-only.
Manually observe the new logging statements are emitted when pairing
fails or does not start.

Change-Id: I3275b587df4307d4c0bca9f4df5581a9480cfd71
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515612
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
Commit-Queue: Xo Wang <xow@google.com>