[bt][gap-bredr] Handle early pairing failures specifically
In PairingState, log and proxy errors received within Simple Pairing
Complete events that are received "early," such as when a peer does not
want to pair because our IO Capability does not meet its requirements.

This was already handled as an "unexpected event" so the difference is
in logging and the error code returned to the pairing client.

Also document these possibilities in the header.

Bug: fxbug.dev/74237
Test: in bt-host-gap-tests,
GAP_PairingStateTest.SimplePairingCompleteWithErrorCodeReceivedEarlyFailsPairing

Change-Id: Id520c757ea6bfbf01dbc367f5b01f1cea3902662
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515615
Commit-Queue: Xo Wang <xow@google.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>