[bt][cleanup] reformat comments
This change simply reformats comments to adjust to the line length of
the file. There are no other logic changes within this change. This
change is introduced mainly as a precursor to larger changes coming up
shortly. The change is intended to avoid noise in the larger changes.

Change-Id: Ie4efeb095e5c5c6643791801bec0aab3c1a73a73
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533531
Commit-Queue: Faraaz Sareshwala <fsareshwala@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>