[bt][hci] add AdvertisingHandleMap
When communicating with the controller, we do so in terms of an
AdvertisingHandle. This means that we frequently need to convert between
a DeviceAddress and an AdvertisingHandle. AdvertisingHandleMap provides
a bidirectional mapping between a DeviceAddress and an
AdvertisingHandle, ensuring a 1:1 mapping between DeviceAddress and
AdvertisingHandle.

Bug: 78081
Change-Id: I25eef4495f5363fd01ef9656fba71231b3ca2a2d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539146
Commit-Queue: Faraaz Sareshwala <fsareshwala@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>