pw_build: Fix exception handling in mirror_tree
When linking fails, we fall back to copying. This always fails because
the original reception is re-raised even if the copy succeeds.

Just remove the exception management in the copy case. If copying fails,
we'll get both exceptions reported in a chain.

No-Docs-Update-Reason: bug fix
Change-Id: I49f0b381ba0e607f46725d1cd59bb16b811c7cf2
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/48700
Commit-Queue: Michael Spang <spang@google.com>
Reviewed-by: Rob Mohr <mohrr@google.com>