pw_env_setup: Add "system_packages" option to the config file
Add a new optional boolean value virtualenv.system_packages
to the environment configuration file, giving the Python
virtualenv access to the system packages. It's achieved by
passing --system-site-packages flag when the virtualenv is
created.

The option is useful when Pigweed is used in the context
of existing Python-based infrastructure such as a test
framework.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>
Change-Id: I137148afcffc0a3ec30426df6dc725fd3eea156e
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/49040
Reviewed-by: Rob Mohr <mohrr@google.com>
Reviewed-by: Michael Spang <spang@google.com>
Commit-Queue: Rob Mohr <mohrr@google.com>