pw_sync_embos: Disable task switching in the ISL
Updates the pw::sync::InterruptSpinLock implementation for embOS to
disable task switching if not in an interrupt. This way there's no
risk that the current task gets switched out to another which
recursively attempts to grab the same lock.

This also changes the internal state to use a plain bool instead
of an atomic bool as an atomic is not necessary.

Change-Id: Ic30bdbee2b298398056d0e5dcdae878ac681f95d
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/50462
Reviewed-by: Wyatt Hepler <hepler@google.com>
Reviewed-by: Keir Mierle <keir@google.com>
Commit-Queue: Auto-Submit <auto-submit@pigweed.google.com.iam.gserviceaccount.com>
Pigweed-Auto-Submit: Ewout van Bekkum <ewout@google.com>