pw_sync_threadx: Disable thread switching in the ISL
Updates the pw::sync::InterruptSpinLock implementation for ThreadX
to disable thread switching if not in an interrupt by raising the
preemption threshold. This way there's no risk that the current
thread gets switched out to another which recursively attempts to
grab the same lock.

This also changes the internal state to use a plain bool instead
of an atomic bool as an atomic is not necessary.

Change-Id: Id02869e66abaee956d1275dc181d1c14777c9cbe
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/50463
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Ewout van Bekkum <ewout@google.com>
Pigweed-Auto-Submit: Ewout van Bekkum <ewout@google.com>