pw_ring_buffer: Add late-reader/drain attach
Changes the default behavior of the ring buffer and multisink such that
late-attach of a Reader or Drain now places the object at the index of
the oldest entry in the buffer.

For Readers, the oldest entry corresponds to the slowest reader present.
If no readers are present, the Reader is placed at the write head. For
Drains, a lagging Reader is always present, so new Drains will always
point to the oldest valid entry in the buffer.

Change-Id: I2b0da0dd7389b98ff121528a0fed38adfa75daf1
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/53344
Commit-Queue: Prashanth Swaminathan <prashanthsw@google.com>
Reviewed-by: Ewout van Bekkum <ewout@google.com>