[bt][sm] Prevent crash 30s after authenticated security request
Upon reception of the Security Request command, we start the SMP timer.
However, if the Security Request asks for authenticated pairing and the
local IOCapabilities don't support authenticated pairing, all we did
before was notify the locally-initiated callbacks of failure, leaving
the peer hanging. This was not the most sensible behavior, because the
Security Request was peer-initiated. It also meant that we didn't stop
the SMP timer, which would cause us to error out 30s later when the
timer expired while no pairing was in progress, as it is an invariant
of the Security Manager that the SMP timer only run while a security
upgrade is in progress.

To fix it, we refactor Security Manager such that the SMP timer is
stopped and the peer is notified of rejection.

Fixed: 78855
Test: `fx test bt-host-sm-tests`, added SMP_InitiatorPairingTest.
      ReceiveMITMSecurityRequestLocalIoCapNoInputNoOutput.
Change-Id: Ia4ac8a79701670f2093677bcbb546a8283c93554
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556241
Commit-Queue: Lucas Jenkins <lucasjenkins@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>