[bt][hci] ensure AdvertisingHandleMap issues AdvertisingHandles in sequential order
There is no guarantee of order or sequence in the values that the
AdvertisingHandleMap returns except that it returns a valid handle in
the range of valid handles and that the handle is currently unused.
However, for the internal consistency of the class and keeping it
clearer to read for future authors, ensure that it returns
AdvertisingHandles in sequential order.

Change-Id: I5f0a1274fe074bac2da8286232ac9527ac27eb4b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/559861
Fuchsia-Auto-Submit: Faraaz Sareshwala <fsareshwala@google.com>
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>