pw_env_setup: Fix pip upgrading on Mac
When bootstrapping with the system python on Mac, __PYVENV_LAUNCHER__
is set in the environment which causes pip installs in the virtualenv
to not work properly. This was fixed for ninja installs but not for
other calls to pip such as the initial upgrade of pip itself.

Fix this by always running pip in the virtualenv during bootstrap.

No-Docs-Update-Reason: bug fix

Change-Id: I21cdae2230671195683f665beb1122018da35d81
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/58600
Reviewed-by: Rob Mohr <mohrr@google.com>
Commit-Queue: Michael Spang <spang@google.com>