pw_rpc: Fix destructor ordering problem in test helpers
During destruction, the FakeChannelOutput& in the InvocationContext base
class would be used to send a final RPC message. The ChannelOutput
instance it pointed to was in a derived class, so would be destructed
before the RPC packet was sent, resulting in "pure virtual function
call" crashes in tests.

- Move the FakeChannelOutput instance to the base InvocationContext so
  it is destructed last, after any responses are sent.
- Explicitly delete some copies/moves.
- Fix typo in transfer_test.cc.
- Handle some ignored Status returns.

Change-Id: I1439dc9af59070507c5dbedab58211f2d165c35d
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/59761
Reviewed-by: Carlos Chinchilla <cachinchilla@google.com>
Reviewed-by: Keir Mierle <keir@google.com>
Commit-Queue: Wyatt Hepler <hepler@google.com>
Pigweed-Auto-Submit: Wyatt Hepler <hepler@google.com>