pw_rpc: Support listening to responses in Python client
Add an open() call to the Python RPC client alongside invoke(). Open
ignores errors when sending the initial RPC request. This can be used to
listen for responses from an RPC that was not explicitly requested. This
is helpful when an RPC server cannot be immediately reached (e.g. the
device isn't plugged in yet).

Change-Id: Id96c840a033b814750fced8c784ea078d8f30aa8
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/59640
Reviewed-by: Carlos Chinchilla <cachinchilla@google.com>
Reviewed-by: Alexei Frolov <frolv@google.com>
Commit-Queue: Wyatt Hepler <hepler@google.com>