pw_rpc: Update C++ client API
This updates the generated service client API to work with an RPC client
and channel ID instead of a channel itself. This will allow removing the
client reference from the channel object.

As part of this, the generated client classes are made instantiable to
improve usability. A client and channel ID must be passed in when
creating a service client; beyond this, they are not needed to invoke
methods.

Change-Id: I0cb29169ccae648b75f6e1d523cca5fe49bd00a1
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/60061
Commit-Queue: Alexei Frolov <frolv@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>