pw_rpc: Replace existing client calls in RegisterCall
This updates the RPC client to replace (cancel) an existing call when a
new one is started for the same RPC instead of failing. This is
consistent with the Python client and expected user behavior (calling an
RPC with the same context should replace the first).

Change-Id: Idb5096a19b83b291cbb714204bc11be5eaa47947
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/60820
Pigweed-Auto-Submit: Alexei Frolov <frolv@google.com>
Commit-Queue: Auto-Submit <auto-submit@pigweed.google.com.iam.gserviceaccount.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>