pw_console: Remove Ctrl-Alt-hjkl and Ctrl-w bindings
Backspace in many terminals is mapped to Ctrl-H. The move window left
keybinding Ctrl-Alt-H clashes with this if you press Alt-Backspace to
delete previous word in the repl. Since Ctrl-Alt-Left also moves the
window left I'm removing these bindings. It's also easy to move windows
using the View menu and the mouse only.

Removed Ctrl-W keybind to quit. This conflicts with vim ctrl-w for
window movement and Chrome ctrl-w to close a tab. Quiting can be done
with the mouse in the File > Exit menu or by typing 'quit' or 'exit' in
the repl.

Bug fix to tab view: If a window pane is hidden then view is switched
to tabbed an exception will occur. This un-hides all panes when
switching to tabbed mode. Also prevents hiding windows if tabbed mode
is enabled.

Change-Id: I538e6cd4b147773b880b92c3aae01592b0a8e278
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/62040
Reviewed-by: Keir Mierle <keir@google.com>
Commit-Queue: Anthony DiGirolamo <tonymd@google.com>