pw_chrono_embos: Always defer callback to the timer ISR
Updates the embOS backend for the SystemTimer to always pass a
non-zero timer period to the native embOS API. This means that
expired deadlines will no longer be executed at the callsite and
may be delayed up to a tick.

This is consistent with the FreeRTOS SystemTimer backend.

Change-Id: I78a4c54c85bc6bec6bcbf6162c9715f70a225191
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/63762
Reviewed-by: Max Koopman <koopman@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Ewout van Bekkum <ewout@google.com>
Pigweed-Auto-Submit: Ewout van Bekkum <ewout@google.com>