pw_transfer: Resend transfer params on client retries
To avoid a flood of transfer parameters updates, the service only sends
a transfer parameters update the first time it receives an incorrect
offset. If that update is dropped, the transfer may stall.

The client retries by resending the last chunk in the window. Have the
service resend transfer parameters when it detects that the client is
retrying.

Change-Id: I0b36cc23b1c0f32b703c8997ade7c4661f013d6e
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/64083
Pigweed-Auto-Submit: Wyatt Hepler <hepler@google.com>
Commit-Queue: Auto-Submit <auto-submit@pigweed.google.com.iam.gserviceaccount.com>
Reviewed-by: Keir Mierle <keir@google.com>