pw_blob_store: Assume not erased if there's no valid KVS entry
Instead of scanning entire flash partitions to verify whether or
not a partition is erased, assume it is not erased and let projects
specialize the FlashMemory and/or FlashPartition implementations to
optimize out erase operations if they'd like.

This removes a very CPU intensive operation out of BlobStore::Init().

Modifies ValidToWrite() to use the flash pointer instead of the
buffer pointer, to ensure that deferred erase can be supported in
a way where flush failures are reported to the deferred writer but
only once flush moves forward.

Also cleans up Flush to use PW_TRY.

Change-Id: I232d3066952fc3196918446833e0f13183e5bd10
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/64085
Reviewed-by: Ewout van Bekkum <ewout@google.com>
Commit-Queue: David Rogers <davidrogers@google.com>