pw_watch: Exit cleanly on max inotify instances
Watching for changes can fail when the max number of instances is
reached. Catch the error and print instructions on how to increase the
max value. This is different from inotify watches which can also fail in
the same way for the same reason.

No-Docs-Update-Reason: bugfix
Change-Id: Ia9df0c21ffcd118b99679476ee8bcebefa91f9ce
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/64802
Commit-Queue: Keir Mierle <keir@google.com>
Reviewed-by: Keir Mierle <keir@google.com>