pw_polyfill: Remove -included language_features.h
The language_features.h header was expected to be included in all source
files with the -include option. It was only adding support for one
feature, static_assert with no message, which is easy to avoid when
required. Removing language_features.h simplifies building Pigweed,
particularly when integrating it with an existing build system, and
avoids some hidden "magic" functionality.

Change-Id: Ibb87ed7da6bbbc1bfc3438b8547a14ede9bcce9a
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/65241
Commit-Queue: Wyatt Hepler <hepler@google.com>
Reviewed-by: Keir Mierle <keir@google.com>