pw_rpc: Add missing pw_log dependency for CMake
When trying to build a simple CMake project with pw_rpc I found that
I was getting errors importing "pw_log/log.h". This change fixes the
error. It seems that this dependency is included in the BUILD.gn files
and was missing in the CMakeLists.txt

Change-Id: Ica05a7f6004a63a8526f27ff9255456151105695
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/65380
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Wyatt Hepler <hepler@google.com>
Commit-Queue: Yuval Peress <peress@google.com>