docs: Display only titles in module guide
Before this change, the autogenerated TOC also included the top-level
headings within the module documents if they were at the same heading
level as the document title.  Example section:

[snip]
*  pw_bloat
*  pw_blob_store
*  Usage
*  FileSystem RPC integration
*  Size report
*  pw_bluetooth_hci
*  pw_boot
[snip]

The "Usage" (etc) are subsections of pw_blob_store docs. Arguably we
should not have any headings in the docs at the same level as the title,
but the simplest and most robust fix is just not to display them here.

Change-Id: Ic32843d7572b11edc14af382e893ee4bef6ea946
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/65724
Reviewed-by: Anthony DiGirolamo <tonymd@google.com>
Commit-Queue: Ted Pudlik <tpudlik@google.com>