pw_rpc: Add call IDs to protocol
This expands the RPC protocol to include an optional call_id field, used
to map server->client packets to the client call object that made them.
This fixes an issue where a reinvoked call could receive a response
intended for the original invocation, with potentially stale data.

This feature is opt-in for RPC clients. Setting all call_ids to the same
number (such as the proto3 default 0) results in the previous client
behavior. Therefore, it is backwards compatible with all existing RPC
clients.

Change-Id: I6f1894afd491e6308dca5c31ae1704af920381c6
Requires: pigweed-internal:16840
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/66064
Commit-Queue: Alexei Frolov <frolv@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>