pw_transfer: Make C++ client work end to end
This updates the pw_transfer C++ client to make it functional. The fake
RPC stubs are removed and replaced with real RPC calls, and unit tests
are added to verify basic client functionality. The client's transfer
logic is mostly removed in favor of the updated Context class.

Change-Id: I0f8f8211115a36b9b74ea33372adbf668b1178c0
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/67231
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Alexei Frolov <frolv@google.com>