pw_transfer: Add progress callback
This allows users to optionally provide a transfer progress callback
when starting a write transfer from the TS client. If provided, the
callback is invoked with the state of the transfer as chunks are
received from the server.

Change-Id: Ib85fabe6ebb4c46e84b42bc261d0eedabd357cb6
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/67660
Commit-Queue: Auto-Submit <auto-submit@pigweed.google.com.iam.gserviceaccount.com>
Pigweed-Auto-Submit: Jared Weinstein <jaredweinstein@google.com>
Reviewed-by: Paul Mathieu <paulmathieu@google.com>