pw_transfer: C++ timeouts and retries
This adds retry support to the C++ pw_transfer logic. A system timer is
added to each Context, which is set whenever a chunk is sent, acting as
a timeout when waiting for a response. If no response is received, the
timeout handler attempts to resend the most recent transfer state.

Change-Id: I8ed923d02f376d13ccf53635b1c13d133904fec9
Requires: pigweed-internal:18040
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/68420
Reviewed-by: Ewout van Bekkum <ewout@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Alexei Frolov <frolv@google.com>