pw_software_update/py: Relax root metadata signing checks
When signing a root metadata, the signing key does not need to be listed
in the root metadata. This CL removes those checks.

Use case: When rotating a root key, the new root metadata need to signed
with both the existing root key and the new root key but the old root
key does NOT need to be listed in the new root metadata. In fact if the
rotation aims to revoke the old root key, the old root key MUST NOT be
listed in the new root signature requirements.

Change-Id: If6896c98ff49cf2dcc936f5d075a93dd440b18a6
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/69360
Reviewed-by: Joe Ethier <jethier@google.com>
Commit-Queue: Ali Zhang <alizhang@google.com>