[bt][common] Remove RunTaskSync utility
Delete the bt::RunTaskSync utility. This is no longer
necessary now that the bt-host stack is single threaded.

Initialize AclDataChannel::io_dispatcher_ in the constructor since
AclDataChannel is initialized in the same thread that it is
constructed on.

Test: fx test bt-host-transport-tests
Change-Id: I19a02f01dad71a269213f69192c74a2e5d75ca41
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608464
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Commit-Queue: Ben Lawson <benlawson@google.com>