cmake: Set the PW_ROOT env variable in pigweed.cmake
Instead of setting the PW_ROOT environment variable inside of the
root CMakeLists.txt, it is set if if needed through
pw_build/pigweed.cmake.

In addition, the provided toolchain files are updated to use the
relative path to ../../pw_build/pigweed.cmake to ensure they work
in environments which are not bootstraped.

This ensures that both the upstream toolchains and Pigweed's
CMake tooling in general can be used outside of a bootstrapped
environment.

Change-Id: I3880ec8a8cb1f9be9c97c1b4df336f211a3aebc2
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/74360
Reviewed-by: Wyatt Hepler <hepler@google.com>
Pigweed-Auto-Submit: Ewout van Bekkum <ewout@google.com>
Commit-Queue: Ewout van Bekkum <ewout@google.com>