pw_protobuf: Support decoding from non-seekable Reader
Allowing a StreamDecoder to operate on a non-seekable Reader means that
protobuf messages can be decoded directly from a SocketStream and other
similar streams.

The previous requirement was to support skipping fields, and advancing
to the end of nested length-delimited fields that were not fully read.
Since this use is always forwards, it can be replaced in those streams
by simply reading and discarding the data.

For the requirement of maintaining the bounds of nested length-delimited
fields, the decoder can maintain the position itself as it reads.

StreamDecoder (and StreamDecoder::BytesReader) can still support
seeking where the underlying stream supports it, updating the
tracked position within the bounds of the field.

Change-Id: Id4d1d41ae0492693bb231732546caba8db871cb3
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/74721
Pigweed-Auto-Submit: Scott James Remnant <keybuk@google.com>
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: Auto-Submit <auto-submit@pigweed.google.com.iam.gserviceaccount.com>