pw_router: Local per-packet parsers
This changes the StaticRouter API to take a PacketParser with each call
to RoutePacket rather than as a member. This PacketParser instance is
then forwarded to router egresses in place of the metadata struct (which
is removed altogether).

This has several benefits:

  1. Different formats of packets can be sent through the same router.
  2. No internal synchronization is required as there are no longer any
     shared resources.
  3. Users can forward custom data to router egresses instead of having
     to use the set of fields defined by the router.

Change-Id: I20ddb9af8b6e2ef71bd6f60f82b688fd4a8010e8
Requires: pigweed-internal:19780
Requires: pigweed-internal:19781
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/75960
Reviewed-by: Armando Montanez <amontanez@google.com>
Commit-Queue: Alexei Frolov <frolv@google.com>