pw_log_string: Add backend
pw_log_string is similar to pw_log_basic in that it simply takes raw
data at logging callsites and pipes it to an extern C symbol. The
difference is that pw_log_string does NOT provide an implementation of
said C symbol. This allows string-based log handling to be dictated by
the project.

Change-Id: Id40262a090a1c17db25425389fe52dce3c546c39
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/77182
Reviewed-by: Carlos Chinchilla <cachinchilla@google.com>
Reviewed-by: Keir Mierle <keir@google.com>
Commit-Queue: Armando Montanez <amontanez@google.com>