pw_transfer: Don't call Finish() on inactive transfers in the C++ client
This adds the same bugfix as pwrev/75980 to the transfer client.

These bugs are indicative of a design issue in the transfer client;
inactive transfers contexts should not be invoked. This will be fixed
in a later change -- a tracking bug is created.

Change-Id: I0ffc45d3c9343587fd4574ce13cd8e2762b612ea
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/77200
Pigweed-Auto-Submit: Alexei Frolov <frolv@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Auto-Submit <auto-submit@pigweed.google.com.iam.gserviceaccount.com>