pw_assert: Assert macro facade
Allows PW_ASSERT/PW_DASSERT macros to expand to macro-based handlers.
This greatly expands the quantity of information that can potentially be
captured at a callsite. The API established by this CL is transitional,
and will see breakages in the future as pw_assert's API is reassessed.

Bug: 246
Change-Id: Id4682e961bff32078e1528b11d126708d463c7f3
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/78188
Reviewed-by: Keir Mierle <keir@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Armando Montanez <amontanez@google.com>