[bt][l2cap] Explicitly memoize FCS computation
Add code to explicitly memoize the "per-octet" subroutine of ComputeFcs
by filling a look-up table at compile time using values from
ComputeRemainderForOctet.

Update some documentation to remove references to logical LSb in the C++
computation model since that is the opposite significance to the modulo
math. Instead just refer to "left" or "right" and only mention LSb and
MSb in reference to the Bluetooth spec.

Fix bug in the way that unused_remainder is computed (the result is
correct but if the previous expression were ported to uint32_t for a
32-bit CRC then it'd produce wrong results, which is counter to the
point of documenting the code).

Bug: 91277
Test: Refactor only. Ran bt-host tests
Change-Id: Id3b30f4e9c0ab3247a93105c90eb6866af437806
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/629721
Reviewed-by: Ben Lawson <benlawson@google.com>
Commit-Queue: Xo Wang <xow@google.com>