pw_{log, log_rpc, system}: Log drop support
Keep a sequence ID for all the logs that will be streamed. Add the
sequence ID of the first log entry in the LogEntries proto. The client
can deduct the logs dropped by the interface based on the sequence ID
and how many LogEntry protos are in a LogEntries proto, as the pw_system
console does. The pw_system console provides a way to report the drop
count.

Change-Id: Icefed676e6671fac75a8be90eaf5178507900b18
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/78740
Reviewed-by: Armando Montanez <amontanez@google.com>
Commit-Queue: Carlos Chinchilla <cachinchilla@google.com>