pw_cpu_exception_cortex_m: Expose LogExceptionAnalysis()
Exposes an internal part of the support backend to log the
exception analysis as an optional utility API.

The cpu_state.cc file is renamed to support.cc to reflect what
the facade/backend that the file is associated with.

The cpu_state.h is pulled out to a separate target to remove
unnecessary circular dependencies within the module.

In addition the last remaining TODOs for pwbug/296 are removed and
the Bazel build files are cleaned up a little bit.

Also updates the docs to include the configuration options.

Bug: 296
Change-Id: Id2ea6adc5dc4f3f99c2d9cf15ea048886739736a
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/78860
Reviewed-by: Armando Montanez <amontanez@google.com>
Commit-Queue: Ewout van Bekkum <ewout@google.com>