pw_protobuf: Correct MaxScratchBufferSize type
It is unusual for a free function in a header to be marked static,
because it tells the compiler to generate a separate function in each
translation unit that includes it
(https://stackoverflow.com/a/22103109). It makes more sense to mark it
for inlining.

Change-Id: Ie8685bf7ae726b38bf9fc050dfdef8c4fdf3e2a6
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/79260
Reviewed-by: Armando Montanez <amontanez@google.com>
Commit-Queue: Ted Pudlik <tpudlik@google.com>