pw_build: Fix Python Tree Tests
There was a problem with this test if the setuptools-scm package was
installed inside the Pigweed virtualenv. An exception gets raised if
the fake pyproject.toml does not have a valid syntax:

  tomli.TOMLDecodeError: Expected "=" after a key in a key/value pair

The exception doesn't occur if setuptools-scm is not installed. It
can't hurt to use valid file contents for these tests so this change
does that.

Change-Id: I555297a21f42dd41bd962db51a556bcb6d42ed6e
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/80222
Reviewed-by: Joe Ethier <jethier@google.com>
Commit-Queue: Anthony DiGirolamo <tonymd@google.com>