pw_containers: Make list iterators safe against deletion
Before this change, if an element is deleted while iterating
over the list, the iteration would never terminate because
the current item's next pointer now points to itself.

This patch adds a shadow copy of the current item's next
pointer to the iterator and uses that when incrementing instead
of the current item's next pointer.

After this change, the following code safe:
    for (auto &item : list) {
      list.remote(item);
    }

Change-Id: I1cd7c051df56e3a8e19a86533c24b1e9f9531c61
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/81447
Reviewed-by: Keir Mierle <keir@google.com>
Commit-Queue: Erik Gilling <konkers@google.com>