[bt][pairing] Signal on responder recalling key
When we are the responder (the peer initiates the pairing request) and
we have a key to provide, the pairing state previously never signaled
completion of pairing, meaning that the "initiating connection" never
was cleared out.

This resulted in the upper-layer state never being "connected" after
the semantic change to require a pairing success.

When we are waiting for encryption to be enabled, it is possible our key
is not the correct key. Allow the peer to re-start the pairing process
in this state.

Bug: b/213331342
Test: added unit tests; fx test bt-host-gap-tests
Test: confirm by connecting from peer and observing "connected" state
Change-Id: Id44c6677df1e04c6e8221908fd713f967a78aec2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631023
Reviewed-by: Ben Lawson <benlawson@google.com>
Fuchsia-Auto-Submit: Marie Janssen <jamuraa@google.com>
Reviewed-by: Xo Wang <xow@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>