third_party/freertos: Always provide pw_third_party.freertos
Updates the CMake build to always provide the
pw_third_party.freertos target similar to how facades work. Unless
configured it is an interface library which depends on an actionable
error command which instructs the user if it is built that it must be
configured first.

This ensures that the CMake build can evaluate the target regardless
of the toolchain and configuration that is being used.

Change-Id: I2709e52603b47f1f1d22c3fe70a665de3978a8d0
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/82760
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Ewout van Bekkum <ewout@google.com>
Pigweed-Auto-Submit: Ewout van Bekkum <ewout@google.com>