pw_log_rpc: Make log entries checks more readable
Use the same order in which entries are added to the MultiSink to verify
the entries received in unit tests. This avoids reversing the order of
the messages, which was needed before. Update unit test to use the more
understandable order.

Change-Id: I2f6a183d25483eebb363572641ca06cb3f2f7f36
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/85460
Reviewed-by: Armando Montanez <amontanez@google.com>
Commit-Queue: Carlos Chinchilla <cachinchilla@google.com>
Pigweed-Auto-Submit: Carlos Chinchilla <cachinchilla@google.com>