pw_doctor: Don't warn mismatched unames on M1 Macs
Since we support M1 Macs via Rosetta, if uname is called from
our environment, it will claim to be running on x86_64. But the
environment was bootstrapped from the system Python, which runs
natively and will yield a uname saying it's running on arm64.
This mismatch is okay and doesn't require any user action, so
let's suppress the warning.

Bug: b/224408077
Change-Id: I8380af2aafe62516c0404f9f4caf7919887c3cc3
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/89020
Reviewed-by: Rob Mohr <mohrr@google.com>
Commit-Queue: Chad Norvell <chadnorvell@google.com>