pw_protobuf: StreamDecoder support for table-based decoding
This implements the C++ side of decoding a protobuf into a atruct
described by a series of MessageField structures.

A new protected StreamDecoder::Read() method is added that accepts a
span of bytes corresponding to a 'struct Message', and a span of
MessageField describing the fields therein.

Change-Id: I3710b47a4e547a274538f647cac18af8086d3896
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/89425
Reviewed-by: Keir Mierle <keir@google.com>
Commit-Queue: Scott James Remnant <keybuk@google.com>