pw_protobuf: StreamEncoder and MemoryEncoder support for tables
This implements protobuf encoding from a struct described by a series of
MessageField structures.

A new protected StreamEncoder::Write() method is added that accepts a
span of bytes corresponding to a 'struct Message', and a span of
MessageField describing the fields therein.

The generated code is extended to add a Write() method to each
StreamEncoder and MemoryEncoder wrapper that accepts a struct of the
correct type and passes in the corresponding fields table.

Change-Id: I7c67a5bbf4920ae5f9b8a561babee58c4dac7e33
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/89641
Reviewed-by: Keir Mierle <keir@google.com>
Commit-Queue: Scott James Remnant <keybuk@google.com>
Pigweed-Auto-Submit: Scott James Remnant <keybuk@google.com>