[bt] Add checks and comments to StaticByteBuffer
StaticByteBuffer's ctor is intentionally flexible to allow different
types to be passed to it for silent conversion, so document the
tradeoffs. Also fix the example in its comment.

Add a check for small inputs to verify that no values provided to it are
losing precision through the conversion to uint8_t.

Bug: 41797
Test: fx test //src/connectivity/bluetooth/core/bt-host
Change-Id: I98d2e296fbc3bce3723ba9e5bfbf4eedf1223a30
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666856
Reviewed-by: Ben Lawson <benlawson@google.com>
Fuchsia-Auto-Submit: Xo Wang <xow@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>