pw_env_setup: Make Rosetta optional on M1 Macs
Projects that use host-provided tooling on M1 Macs (i.e. the Xcode
toolchain) need, at minimum, the mac-arm64 version of GN. So we can't
indiscriminately fall back to mac-amd64 binaries.

This solution allows the user to tell bootstrap.sh whether they want to
use Rosetta or not. If not, they will currently get an incomplete CIPD
toolchain, which will prevent upstream development but might be fine for
some downstream projects that don't need the entire toolchain. If the
user instead chooses to use Rosetta (the default), they will get
mac-amd64 binaries for the entire toolchain and everything will work as
it currently does.

An alternate solution to this one would be to get mac-arm64 binaries
when they're available, and only fall back to mac-amd64 when they're
not. Whether it's done manually (by curating the CIPD setup JSON files)
or automatically (by looking at what's available on CIPD at the time of
bootstrapping), this approach was deemed to be too error-prone and
confusing. The balance between mac-arm64 and mac-amd64 binaries could
shift regularly based on CIPD package availability, causing
difficult-to-debug incompatibilities between different parts of the
toolchain, and different developers on the same project wouldn't be
guaranteed to have the same toolchain. It's better to set the
expectation that the Pigweed toolchain will be of exactly one
architecture.

Bug: b/228610348
Change-Id: Idd1bf13ddf330281180907ccb69dde17f1b3a679
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/90380
Reviewed-by: Ted Pudlik <tpudlik@google.com>
Commit-Queue: Chad Norvell <chadnorvell@google.com>