pw_hdlc: Make Python impl retain flag bytes
If the HDLC flag bytes are discarded, users of the library can not
do lossless processing of error frames.  This causes more frame
errors to be returned and the tests are updated accordingly.
Since the C++ and TS libraries are not updated by this change,
these frame errors are filtered out in the generation of their
tests.

Change-Id: Ifba9dfe3272ab7645648169f5c57bd802fb09e52
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/90941
Reviewed-by: Ted Pudlik <tpudlik@google.com>
Commit-Queue: Erik Gilling <konkers@google.com>