third_party/boringssl: Document how to include this module
boringssl is a bit special since the BUILD.gn file includes a .gni from
the externally provided source directory. This means that if the path
to the source code (dir_third_party_boringssl) is defined in GN as a
default value the code must be present during bootstrap since GN needs
to parse the .gni file. In other third_party modules this is not a
problem since the source code is only used at compile time. In addition
to this, the .gni file is a generated one, not present in upstream
boringssl repository. "pw package" install process generates the file
but only after the initial system is installed.

This patch adds a docs.rst documentation explaining how to set up
boringssl with "pw package" avoiding these issues.

Change-Id: I94d2441400f05d9cc1ae1781e38af6ba0d8b3d74
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/92120
Reviewed-by: Rob Mohr <mohrr@google.com>
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: Alex Deymo <deymo@google.com>