pw_stream: Remove use of std::tmpnam from test
Removes usage of std::tmpnam() and replaces it with a custom
cross-platform way to create temporary files.

While std::tmpnam() is part of the standard library, both Linux and
macOS mark the function as deprecated and emit a warning saying not to
use it and that the function only continues to exist for compliance
reasons.

Change-Id: I8e82db1ce657ddc7b36469f9e456134b2f4be6d9
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/93801
Commit-Queue: Auto-Submit <auto-submit@pigweed.google.com.iam.gserviceaccount.com>
Pigweed-Auto-Submit: Armando Montanez <amontanez@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>