pw_transfer: Add separate session_id to chunk
This partially reverts pigweed:89260 by renaming field number 1 in the
chunk proto back to transfer_id. The session_id field is introduced as
an entirely new field, so that its presence can be used to detect the
protocol version that a transfer is running.

Change-Id: I9765edc6fa5a263de1366176c2e08e8e5a1bd0f0
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/94063
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Alexei Frolov <frolv@google.com>