[bt][gap] Do not print invalid UTF8 peer names
When a peer name is registered that is not valid UTF8, do not log the
name (just log a warning). The name may contain PII, and the way the
name is currently printed uses sstream, which is a huge dependency that
we want to avoid.

Test: fx test bt-host-gap-tests
Change-Id: I1bf4d6034378cecc2196de904bc8f4d875ad2dab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679447
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Ben Lawson <benlawson@google.com>