[bt][gap] Use SequentialCommandRunner in LowEnergyInterrogator
Refactor LowEnergyInterrogator to use SequentialCommandRunner
to send commands. This greatly simplifies the code and removes
usage of fbl::RefPtr.

Refactor LowEnergyInterrogator to correspond to a single peer,
since it is only used to interrogate a single peer in practice.

Safely handle destruction in SequentialCommandRunner::Cancel()
when the status callback is called (this happens when
LowEnergyConnectionManager cancels interrogation).

Remove LowEnergyInterrogator from the Interrogator inheritance
hierarchy. The code is easier to understand now, and changes to
LowEnergyInterrogator can now be decoupled from BrEdrInterrogator.

Bug: 100658
Test: fx test bt-host-gap-tests bt-host-hci-tests

Change-Id: Id7dc62a270b72d4d47b6a08105c4ad7cf406106a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681446
Commit-Queue: Ben Lawson <benlawson@google.com>
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>
Reviewed-by: Faraaz Sareshwala <fsareshwala@google.com>