docs: Add using-directive exception for literals
The Google external style guide forbids the usage of using-directives,
regardless of scope. In Pigweed, there are many using-directives for the
purposes of including literal operators, such as std::chrono_literals
and pw::bytes. This CL adds an explicit exception for such uses to the
rule for the Pigweed codebase.

Change-Id: Ia8c002b9646c400463ab578a8ab2988a0ee34a5e
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/96901
Commit-Queue: RJ Ascani <rjascani@google.com>
Reviewed-by: Armando Montanez <amontanez@google.com>