pw_rpc: Add total_payloads to FakeChannelOutput
FakeChannelOutput.payloads provides a view to the payloads seen for
given RPC, but it can't be used when the same Output is written to by
other threads. total_payloads() method is a thread-safe shortcut that
returns only the number of seen payloads and can be used in tests where
polling for the specific RPC payload is needed.

Change-Id: I3e65b015661c164aa92745780eee0d9011371fe6
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/97561
Pigweed-Auto-Submit: Dennis Kormalev <denk@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Auto-Submit <auto-submit@pigweed.google.com.iam.gserviceaccount.com>