pw_unit_test: Require 'light' backend for RPC service
The pw_unit_test RPC service only supports the 'light' backend. If it's
used with another backend such as GoogleTest, there will be misleading
compilation errors and the service won't find any of the tests.

This commit adds a build assertion to check that the correct backend is
in use if the RPC service is used.

Change-Id: I184c7476d9f88432cc98a686f1f4f75acd6b0a58
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/97907
Commit-Queue: Eli Lipsitz <elipsitz@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>