pw_toolchain: Fix clang-tidy.py multi-arg support
There is was an assumption that after the -- in the clang-tidy
command there would be exactly one word and that would be the
clang binary name.  This prevented having anything else in the
string, including things necessary for RBE.  This change,
adds a marker 'END_OF_INVOKER' to the end of the compile
argument so it can explicitly be checked for so additional
arguments can be added before or after the compile binary.

Bug: None
Test: Verified this generated the build files correct and built
      correctly with RBE and regular builds.

Change-Id: I93a9b608c16d2dc369a27f6ed9e2cdb36aa3cc9f
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/98360
Reviewed-by: Joe Brennan <jmbrenna@google.com>
Reviewed-by: Ted Pudlik <tpudlik@google.com>
Commit-Queue: Auto-Submit <auto-submit@pigweed.google.com.iam.gserviceaccount.com>
Pigweed-Auto-Submit: Joe Brennan <jmbrenna@google.com>