pw_status: Use switch for pw_StatusString
pw_StatusString was switched to a series of if statements rather than a
switch to compile with -Wswitch-enum. Unfortunately, some compilers
(including arm-none-eabi-gcc 10.3 and earlier) do not optimize the if
version of the function very well, resulting in a substantial code size
increase.

This function replaces the if statements with a swtich. It includes the
internal-use-only enum case for -Wswitch-enum compatibility.

Change-Id: I81c0f93c5eb0ba4a3ca935bb97ea2b462766ee97
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/100467
Reviewed-by: Armando Montanez <amontanez@google.com>
Pigweed-Auto-Submit: Wyatt Hepler <hepler@google.com>
Commit-Queue: Auto-Submit <auto-submit@pigweed.google.com.iam.gserviceaccount.com>