pw_kvs: Trim more code when deleted key cleanup is disabled
In addition to removing calls to RemoveDeletedKeyEntries(), remove the
call to HeavyMaintenance() in WriteEntryForNewKey() when deleted key
cleanup is disabled. The HeavyMaintenance() call doesn't accomplish
anything when deleted key cleanup is disabled, but it unconditionally
pulls in the HeavyMaintenance() function. That causes a large code size
increase if the maintenance functions are not already in use.

Change-Id: I7e6cfb9414d1baaa6a5547bf5350d919c87c4e16
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/102142
Commit-Queue: Auto-Submit <auto-submit@pigweed.google.com.iam.gserviceaccount.com>
Pigweed-Auto-Submit: Wyatt Hepler <hepler@google.com>
Reviewed-by: Anthony DiGirolamo <tonymd@google.com>