pw_web_ui: Bundle pw_protobuf_compiler and its template
Currently, we need ts-node to run pw_protobuf_compiler and the template
for proto collection is also in TypeScript which means when user runs
this command on their own protos, the output ProtoCollection is an
unbundled .ts file with missing imports (since we don't include .ts
files in our final npm bundle). This bundles the template as a .js file.

This also installs specific `protoc` binary within the node_modules
folder that works with pw_protobuf_compiler. This means we don't need
user to install the specific version and add it to PATH before running
pw_protobuf_compiler.

Change-Id: I960bdea8d8361fd8ec3fcf6b7830aa264f6bc16a
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/102323
Reviewed-by: Anthony DiGirolamo <tonymd@google.com>
Commit-Queue: Asad Memon <asadmemon@google.com>