pw_web_ui: Replace '@pigweed' and 'pigweed' imports with 'pigweedjs'
Changing package name from pigweed to pigweedjs caused a few protos
to not be included in final bundle. This fixes the bug and adds
a test to check if the bundle imports some outside module, which
happens when rollup fails to find the right module when creating
the bundle.

Change-Id: I0987ca74e91b1b355779bf382ea8b567e62eda18
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/103400
Commit-Queue: Asad Memon <asadmemon@google.com>
Reviewed-by: Anthony DiGirolamo <tonymd@google.com>