pw_build: Add ability to exec multi argument command
There is an assumption in exec.py that the first
argument will always be exactly 1 command with no
additional arguments.  However for RBE this becomes
a multi argument string and when that multi argument
string is fed into subprocess in exec.py, subprocess
crashes.  This checks to see if there are any spaces
in the command splits the command into a list so
subprocess can handle it correctly whether there is
a space or not.

Bug: b/221874708
Test: Verified this generated the build files correct and built
      correctly with RBE and regular builds.
Change-Id: I6521ea03f04a2755422db84f50884e9868796925
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/98442
Reviewed-by: Ted Pudlik <tpudlik@google.com>
Commit-Queue: Joe Brennan <jmbrenna@google.com>
Pigweed-Auto-Submit: Joe Brennan <jmbrenna@google.com>
Reviewed-by: Joe Brennan <jmbrenna@google.com>