pw_unit_test: Export missing types from Python module
The `EventHandler` type, which is exported from the pw_unit_test Python
module, has functions to override that take the TestExpectation
type as arguments. However, the TestExpectation type is not exposed in
the module's __init__.py (unlike other related types), so it's not
easily accessible from dependent code. This fixes the issue.

Additionally, EventHandler has a function that takes the result of a
test case. This is passed as an `int`, using the value of the generated
internal protobuf. To use this in dependent code, you'd need to either
find and use the generated protobuf (which isn't clear because the type
isn't listed), or use magic constants in the code. This commit also
addresses this issue by wrapping these results in an enum
TestCaseResult and exporting that.

Change-Id: I111e486724f89c67e7bce069243a9461948ccac9
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/103486
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Eli Lipsitz <elipsitz@google.com>