pw_rpc: Add PacketMeta
PacketMeta allows users of pw_rpc to examine information about a
packet prior to processing it fully with ProcessPacket. This allows
packets targeting different channels or services to be filtered and
processed within different contexts.

A key motivation of this is to enable users to run ProcessPacket for
certain service+channels on a specific thread, a new thread, or a work
queue, rather than processing all requests on a single main RPC thread.

Change-Id: I3e3ded876740f0561ca79457185693268970708a
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/103960
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Taylor Cramer <cramertj@google.com>