pw_result: Add and_then monadic operation
This CL adds a monadic operation to pw::Result<T> for `and_then`. This
method takes a function, which will be invoked only if the pw::Result<T>
contains a value. If the pw::Result<T> contains a status, then that
status will be returned. This is useful for chaining multiple operations
that might fail.

  Result<std::string> GetIntegerString();
  Result<int> ConvertStringToInteger(const std::string&);

  Result<int> x = GetIntegerString()
                    .and_then(ConvertStringToInteger)
                    .and_then([](int i) { return Result<int>(i * 2) });

Bug: b/236160902
Change-Id: Iafe7cd1c83c2b41f3a58eb78bfd54be32f1746eb
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/100802
Commit-Queue: RJ Ascani <rjascani@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>