pw_result: Add or_else monadic operation
This CL adds a monadic operation to pw::Result<T> for `or_else`. This
method takes a function, which will be invoked only if the pw::Result<T>
contains a status. If the pw::Result<T> contains a value, then that
value will be returned. This is useful for handling error scenarios.

  Result<std::string> GetIntegerString();

  Result<int> x =
    GetIntegerString()
      .or_else([](Status s) { PW_LOG_ERROR("%d", s.code()); });

Bug: b/236160902

Change-Id: If58c5fcdf17ce9b5104c34a13a9aadf82e2ac250
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/100820
Commit-Queue: RJ Ascani <rjascani@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>