pw_result: Add transform monadic operation
This CL adds a monadic operation to pw::Result<T> for `transform`. This
method takes a function, which will be invoked only if the pw::Result<T>
contains a value. If the pw::Result<T> contains a status, then that
status will be returned. This is useful for applying a function to a
Result and transforming the type.

 Result<int> ConvertStringToInteger(std::string_view);
 int MultiplyByTwo(int x) { return x * 2; }

 Result<int> x = ConvertStringToInteger("42")
                   .transform(MultiplyByTwo);

Bug: b/236160902
Change-Id: Iacea9fb7c7482425c30405606531aa90e60c45a7
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/101348
Commit-Queue: RJ Ascani <rjascani@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>