pw_build: Support -include args in pw_linker_script()
Add a new "includes" argument to the pw_linker_script() template to
include from the command line a list of files.

There was a missing "invoker." when using cflags that was causing it to
fail if cflags was passed to pw_linker_script(), which is also fixed
here.

Finally, the invoker.inputs list was only being passed to the
pw_source_set created, which would re-link files that depended on this
linker script, but wouldn't re-generate the final linker script by
running the preprocessor, so files listed as "-include" in cflags or
as #include directives would not work. This patch moves the inputs
argument to the pw_exec so changes there also trigger a re-generation of
the final linker script.

Change-Id: Iab0a03e5b3d23f4e74ad9ef2bdefd5f31c87faf3
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/106110
Reviewed-by: Armando Montanez <amontanez@google.com>
Commit-Queue: Alex Deymo <deymo@google.com>