pw_build: Don't mutate deps list in bazel macros
The deps list is passed by reference. If the caller uses a variable as
the deps list, rather than a literal, that variable will be mutated as
a side effect of the macro invocation if we use `append`. This is not
intended.

Change-Id: Iaa3e1e42b72e809e853eef62252486cdc22b78a7
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/106912
Reviewed-by: Keir Mierle <keir@google.com>
Commit-Queue: Auto-Submit <auto-submit@pigweed.google.com.iam.gserviceaccount.com>
Pigweed-Auto-Submit: Ted Pudlik <tpudlik@google.com>