pw_unit_test: Cleanup test processes properly on timeout
When processes launched were killed due to timeout, their child
processes were orphaned and left running. Add a trap to the test runner
script to cleanup the test processes when terminated. Switch to using
SIGTERM instead of SIGKILL to allow the script to catch and handle it.

Change-Id: Ida60ae12a1c43c2701510a58ab56cba0629ae337
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/107777
Reviewed-by: Ted Pudlik <tpudlik@google.com>
Commit-Queue: Austin Foxley <afoxley@google.com>