pw_sync: add class InlineBorrowable
InlineBorrowable owns an object and a mutex that guards the object.
Both the object and mutex are constructed in-place, and the object can
only be accessed by calling acquire() or one of the acquire_*() methods.

InlineBorrowable should be used whenever an object needs to be guarded
for its entire lifetime by a single mutex. InlineBorrowable extends
Borrowable, so it can be passed to existing APIs by reference.

Change-Id: Ib6f2a0471cf609af10524da7acc8a7de18f2874c
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/103180
Commit-Queue: Anton Markov <amarkov@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>