pw_ide: clangd wrapper generators
clangd must be run within the activated Pigweed environment to pick up
the right paths to the Pigweed toolchains. We point clangd language
servers to these wrapper scripts instead of the bare executable.

We can't just store platform-specific scripts as files in this repo,
because most editors don't let us specify a different clangd location
for different OS's (e.g. `clangd.bat` on Windows and `clangd.sh`
elsewhere). Also, the location of the Pigweed environment can vary. So
we generate wrapper scripts for the user that are appropriate to their
OS and project configuration.

Change-Id: I70a8d3536678329a802507ca950ca79a274aa36f
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/110254
Reviewed-by: Anthony DiGirolamo <tonymd@google.com>
Commit-Queue: Chad Norvell <chadnorvell@google.com>