pw_ide: Create Python symlinks
Provide commands for creating symlinks to the Python virtual environment
and associated bin directory. The location can vary depending on the
name of the environment directory, and the Python venv directory
structure is different on Windows compared to other platforms. This
allows us to provide stable references to the venv to IDEs regardless of
those variations.

Change-Id: I738e9729b695982b2d6a6fe3ae73923efa54d891
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/110256
Reviewed-by: Anthony DiGirolamo <tonymd@google.com>
Commit-Queue: Chad Norvell <chadnorvell@google.com>