pw_bluetooth: Replace pw::Vector with pw::span on input parameters
When passing a struct with a set of parameters such as the
LocalServerInfo or AdvertisingData, use a pw::span<const T> instead of a
pw::Vector<T>. The latter can't be directly constructed unless a variant
with a fixed maximum size is used.

Fixed some spelling typos.

Change-Id: Ic012fb9086097fa077b20702f57049150cbbfdbe
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/112270
Reviewed-by: Ben Lawson <benlawson@google.com>
Commit-Queue: Alex Deymo <deymo@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>