pw_transfer: Work around a GCC 11 to 12.1 bug
GCC 11-12.1 have a bug that causes some valid code initializing
unions in a value initializer to be rejected. However, adding
explicit types causes the code to compile. This commit adds types
to a few places that need them in this module to compile with
these versions of GCC.

This bug appears to be fixed with GCC 12.2, and is not present in
any versions of Clang.

Change-Id: I3b93ea3058925fea275eb8a132f08ef6cdbe12db
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/112864
Pigweed-Auto-Submit: Eli Lipsitz <elipsitz@google.com>
Commit-Queue: Eli Lipsitz <elipsitz@google.com>
Reviewed-by: Ted Pudlik <tpudlik@google.com>