[bt][core] fix std::vector usage in profile_server's DataElementToFidl
A bug was introduced in fxrev.dev/642344 where the incorrect std::vector
constructor is used to copy a string. The code created an std::vector of
the correct size but filled it with the first byte of the source data.
Instead, we need to copy all bytes in sequence over to the new
container.

Fixed: 110224
Test: fx test bt-host-fidl-tests
Change-Id: I51a4c60cc2c2ca52d79d4550285464e9767fb688
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729210
Reviewed-by: Ali Saeed <saeedali@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>
Commit-Queue: Faraaz Sareshwala <fsareshwala@google.com>
Reviewed-by: Matthew Kehrt <kehrt@google.com>