pw_bluetooth: Support non-trivial destructor in pw::bluetooth::Result
pw::bluetooth::Result<E, T> is mostly used in combination with T =
std::unique_ptr<U> which obviously has a non-trivial destructor. This
patch places T in an std::optional<T> instead so we only destroy T when
the result is a non-error condition.

The internal storage size could be optimized to remove the error E
storage on success. This is a simpler temporary solution to make the API
work until this type can be replaced with pw::Result from pw_result.

Test: Added unittests to check non-trivial destructor case.
Change-Id: I1237878f95e7e7fb5cf9e8dfd6f5dbb6561dc1a8
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/112930
Reviewed-by: Ben Lawson <benlawson@google.com>
Reviewed-by: Armando Montanez <amontanez@google.com>
Commit-Queue: Alex Deymo <deymo@google.com>