pw_transfer: Support protocol v2 in Python client
This updates the Python transfer client to work with either the legacy
or version 2 transfer protocols.

To support this, a wrapper class around the generated protobuf Chunk
class is implemented, supporting serializing and deserialzing specific
sets of fields based on the protocol version in use and type of chunk.

To set the protocol version, a default can be provided when constructing
a transfer manager. Alternatively, version can be set on a per-transfer
basis when starting a new transfer. As with the C++ client, the default
version remains LEGACY until all transfer implementations are ready to
switch over.

Bug: 243679452
Change-Id: I30dbca639a3e65ace9afcf3640fa569bf9d8e279
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/111330
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Alexei Frolov <frolv@google.com>