pw_bluetooth: Use pw::Function<T>&& when storing callbacks
When passing callbacks we should use `pw::Function<T>&&` for callbacks
that are meant to be stored by the callee (the common case in
pw_bluetooth) or `const pw::Function<T>&` when the callback is meant to
be called directly from the callee but not copied.

This patch changes all uses of `pw::Function<T>` passed by copy to
`pw::Function<T>&&` to avoid forcing a copy of the passed callback which
for some callbacks is not available.

Change-Id: I698fd69d86ea727f87de385be135afe2161e216a
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/112931
Reviewed-by: Armando Montanez <amontanez@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>
Commit-Queue: Alex Deymo <deymo@google.com>