pw_transfer: Use correct chunk ID when resource doesn't exist
In transfer protocol v2, session IDs are assigned by the server. When a
client sends its opening chunk, it does not yet know its session ID and
expects a response that is identified by the resource ID it is trying to
transfer.

In response to an opening chunk from the client, the server looks up a
handler for the requested resource. If one is not found, the transfer
thread responds directly without assigning a context for the transfer.
This direct response did not account for the initial session/resource ID
discrepancy, instead sending its NOT_FOUND chunk with the newly assigned
session ID of which the client had no knowledge.

This updates the transfer thread to always use resource ID to identify
an error in response to an opening chunk. In legacy protocol, the two
are the same, whereas in v2, resource ID is expected.

Bug: 252870476
Change-Id: Ia594286976e263b6a46cf0bc582a848122386471
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/113990
Commit-Queue: Alexei Frolov <frolv@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>