[bt][common] Add pw::span interop to ByteBuffers
Add ByteBuffer::subspan() method that mirrors the existing view() method
but returns a span instead. This is a very common operation when
converting a packet to a span for use with pw::bluetooth::Controller.

Add BufferView::BufferView(pw::span) constructor for creating a
BufferView from a pw::span. This is a very common operation when
creating pw::bluetooth::Controller backends, such as FakeController.

Bug: 100593
Test: fx test bt-host-common-tests

Change-Id: I9dee2c7486b84e64763833f2ac121cb313f0bf88
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739650
Reviewed-by: Faraaz Sareshwala <fsareshwala@google.com>
Commit-Queue: Ben Lawson <benlawson@google.com>