pw_transfer: Add lifetime retry limit
This adds a lifetime retry limit to the C++ and Python transfer
implementations. If the total number of retries over the duration
of the transfer exceeds this limit, the transfer is canceled. This
is done to avoid situtations where transfer could end up in an
infinite retry loop.

Change-Id: I8572a84a323e1517b7d422e6a19d094c03a88974
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/114490
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Alexei Frolov <frolv@google.com>