pw_bluetooth: Specify the Deleter when using std::unique_ptr
In several cases we use std::unique_ptr<T> to return an object with RAII
semantics, meaning that some action is needed by the backend to release
the resource when the object goes out of scope. However, using using
std::unique_ptr<T> with the default deleter forces the pw_bluetooth
backend to use new/delete to allocate the T interface implementation.

This patch introduces a RaiiPtr<Api, void (Api::*OnDestroy)()> template
class as an alias of std::unique_ptr with a custom deleter that calls
the Api::OnDestroy parametric method instead of calling "delete". This
template class is used instead of the plain std::unique_ptr when
returning objects from the pw_bluetooth API for the all the "Api"
classes with only pure virtual methods.

In all 5 cases where this pattern was used the abstract class has one
more private pure virtual method (the one used as Api::OnDestroy
parameter) which must be implemented by the backend.

This allows a backend to manage the memory for the returned objects with
something other than new/delete, for example it can return a RaiiPtr to
a std::optional instance or to one of the existing pre-allocated objects
and re-use them  when needed. This is particularly useful in cases where
there can only be a fixed small number of objects for the given type,
like in the Scan or AdvertisingParameters case where there is only one
object at a time.

Bug: 250930246
Change-Id: I724ff5135e0f73f5fb74c7fc92abdffa6a9e2d2d
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/113210
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Alex Deymo <deymo@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>