[bt][socket] Fix SocketChannelRelay::OnSocketWritable() crash
Fix SocketChannelRelay crash caused by a failing
`!socket_write_queue_.empty()` assert in
SocketChannelRelay::OnSocketWritable(). This should not be an assert
because there is an edge case where the write queue may be emptied
before OnSocketWritable() is called. In OnChannelDataReceived(), if the
first packet in the write queue gets dropped and the subsequent smaller
packets all fit into the socket, the write queue will be emptied.

Cancel the socket writable wait in OnChannelDataReceived() if the packet
its threshold corresponds to is dropped.

Fixed: 77407
Test: fx test bt-host-socket-tests

Change-Id: I09d290b4d082c777fcc85c32eb39f2d2fc55e2f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/749182
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Ben Lawson <benlawson@google.com>