pw_software_update: Handle `pw update create-root-metadata`
Added the parser and handler for `create-root-metadata`. This
subcommand takes root keys, target keys, outpath, and an optional
version number to create the serialized metadata file.

The handler manages the succcessful creation of the root metadata
and provides the user with "success state" output.

Bug: 254524996
Change-Id: I57567852820399ad5782c610da5c29f4e876c380
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/116290
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: Eashan Singh <eashansingh@google.com>
Reviewed-by: Ali Zhang <alizhang@google.com>