pw_presubmit: Allow reuse of filtered files
When a presubmit program is run in CQ, it's first run with
'--only-list-steps'. Recipe code parses the results of this, which shows
which steps are to be run based on the files changed. Then the steps are
each run individually. This makes the UI of CQ better, because each step
shows separately. However, we don't need to reevaluate the changed files
for each step.

Change the behavior of '--only-list-steps' to also output the list of
files that should be passed into each step. Then add a
'--list-steps-file' argument that reads in this list of files rather
than recalculating it from the checkout on disk.

Since recipe code does not yet use '--list-steps-file', this change has
no immediate effect except to add a list of paths to the output of
'pw presubmit --only-list-steps', which recipe code will ignore.

Bug: b/256007641
Change-Id: I6b32e013e2bdfa22b5637f7551827129a6e5a314
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/116576
Commit-Queue: Auto-Submit <auto-submit@pigweed.google.com.iam.gserviceaccount.com>
Pigweed-Auto-Submit: Rob Mohr <mohrr@google.com>
Reviewed-by: Ted Pudlik <tpudlik@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>