pw_multisink: Fix drop count after PeekEntry()
When removing messages from a Drain with PeekEntry() and then PopEntry()
of the peeked_entry, it is possible that the multisink receives new
messages between two causing the drain to drop more messages (including
the peeked one). This is not a problem for the peeked entry since the
contents are copied, but on PopEntry() the old message is just ignored
and the drop count not updated.

This patch updates the drain's `last_handled_sequence_id_` in this case
so a subsequent PeekEntry() returns the right number of dropped messages
since the last call to PopEntry().

Test: Extended a unittest to cover this situation.
Change-Id: Iafd8e62041ca12b84e381e9ed15f970ad1a52766
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/117750
Reviewed-by: Carlos Chinchilla <cachinchilla@google.com>
Reviewed-by: Ewout van Bekkum <ewout@google.com>
Commit-Queue: Alex Deymo <deymo@google.com>