[bt][l2cap] Stop calling ChannelImpl::rx_cb_ after SignalLinkError()
Stop calling ChannelImpl::rx_cb_ if Channel::SignalLinkError() is called
inside rx_cb_. SignalLinkError() destroys the channel and notifies the
activator of the channel of the closure. Future rx_cb_ calls are
likely unexpected by the client and are likely to result in a crash.

Fixed: 113458
Test: fx test bt-host-l2cap-tests
Change-Id: I683778b6dbe9073ee98f1e2631a0b0628d9fecfc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753806
Reviewed-by: Ali Saeed <saeedali@google.com>
Commit-Queue: Ben Lawson <benlawson@google.com>