pw_log_basic: Make the maximum entry size configurable
The size of the encoded entry depends on the message in the log entry
but also on other pw_log_basic configuration parameters such as whether
the file, line and timestamp are included. The maximum size set in the
code was 150 bytes, which this patch this size is configurable via the
PW_LOG_BASIC_ENTRY_SIZE macro.

Change-Id: I4802a2fa852baca1a605842cdd4dc946185a8231
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/117751
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Alex Deymo <deymo@google.com>