pw_software_update: Create empty bundle CLI subcommand
Add functionality for creation of an empty update bundle(parser +
handler). An empty bundle is created via the `pw update
create-empty-bundle` command.

The 'path' to the empty bundle is a required positional argument.

The target metadata version is an optional argument with default set to
1.

Change-Id: I670914c8b44ad53b014375d37d9ab3aa5d9fafbb
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/117211
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: Eashan Singh <eashansingh@google.com>
Reviewed-by: Ali Zhang <alizhang@google.com>