pw_toolchain: Add host_clang coverage toolchain
Adds a coverage toolchain, similar to the existing asan, msan, tsan,
etc toolchains, that allows creation of concrete toolchains that does
not depend on the state of `pw_toolchain_SANITIZERS` to enable coverage
instrumentation.

This came about in a downstream project when it was hard to make a
coverage toolchain reusing the logic from `host_debug` since
`pw_toolchain_SANITIZERS` could not be set until after the
`pw_toolchain_host_clang` scope would have been interpreted. Instead of
doing this logic in the downstream project, we can do it in Pigweed,
which makes it easier to ensure it will be properly updated with the
others if the pattern changes.

Bug: b/254891135
Change-Id: Ie0e6fd72b3107755182d276b3c3f70e0c57380e4
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/119633
Commit-Queue: Greg Pataky <gregpataky@google.com>
Reviewed-by: Anthony DiGirolamo <tonymd@google.com>