pw_protobuf_compiler: Impl. some common Bazel rule attrs
These are sourced from
https://bazel.build/reference/be/common-definitions#common-attributes. I
did not add everything that was recommended there, but I did add
`visibility` and `tags`, which are common in large/mature Bazel
projects, and can be difficult to do without if the project is
accustomed to using them.

I added a few of them to pw_proto_library. For now, they are plumbed
through to all the sub-rules that pw_proto_library generates. Note that
this rule will eventually be broken up into several, per b/241313480.

Change-Id: I4d47e118ac6f5135c500037698b43b85e5c9416c
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/120410
Commit-Queue: Anqi Dong <anqid@google.com>
Reviewed-by: Ted Pudlik <tpudlik@google.com>