pw_sync_freertos: Suspend Scheduler Within ISL
The FreeRTOS backend implements ISL using an interrupt-disabled
critical section. This may be problematic in the event the
holder of the lock calls a kernel API that results in a higher
priority task becoming unblocked, based on the target-specific
implementation of portYIELD.

If the target port implements portYIELD synchronously, the ready
task will immediately preempt the current-running task - despite
being within a critical section.  As part of the context switch,
the interrupts may also be re-enabled.  These side effects
break the expected behavior of the lock.

This condition may not occur in the event portYIELD is implemented
asynchronously however - as the interrupt required to service
the resulting context switch may be masked for the duration of
the critical section.

This CL modifies the ISL implementation of the ISL to additionally
call vTaskSuspendAll()/xTaskResumeAll() on lock/unlock operations.
This fixes the errant condition, as FreeRTOS ensures that the
current-running task will not be preempted while the scheduler
is suspended.

Change-Id: I8ad75ddec63281a406825b938ae988e4bf9fd92a
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/119810
Reviewed-by: Ewout van Bekkum <ewout@google.com>
Commit-Queue: Mark Slevinsky <markslevinsky@google.com>