[bt][transport] Create EmbossCommandPacketT, which remembers view type
Create EmbossCommandPacketT, a subclass of EmbossCommandPacket that
is templated on the view type. EmbossCommandPacket::New returns
an EmbossCommandPacketT so that clients only need to specify the
view type once. Because EmbossCommandPacketT has no data, it is safe
to slice a EmbossCommandPacketT into an EmbossCommandPacket (e.g.
when calling CommandChannel::SendCommand).

Bug: 86811
Test: fx test //src/connectivity/bluetooth/core/bt-host
Change-Id: Iefe399f086b8543477e9ddd5d1e7aa2fe5cb132c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769085
Reviewed-by: Faraaz Sareshwala <fsareshwala@google.com>
Commit-Queue: Ben Lawson <benlawson@google.com>
Reviewed-by: Ali Saeed <saeedali@google.com>