pw_build: Add pw-wrap-ninja helper
This commit adds a standalone Python script that wraps Ninja to provide
a more comprehensive display of the progress of the build, along with a
brief summary of the tool to the pw_build docs.

The tool works by parsing Ninja's output to build a representation of
the build status, and then outputting a richer TUI with the list of
running targets and how long they're taking. This allows one to, at a
glance, determine what targets are blocking the rest of the build or
taking an unusually long time to complete.

Change-Id: If33d6d322caeecd9f70f8b7d04dcb97f84eb1c81
Bug: 263976434
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/125551
Reviewed-by: Rob Mohr <mohrr@google.com>
Commit-Queue: Eli Lipsitz <elipsitz@google.com>