pw_build: Add argument parsing to pw-wrap-ninja
This commit adds argument parsing to the Ninja wrapper script, to allow
the user to override the UI update rate and the maximum number of Ninja
actions displayed.

It uses argparse.parse_known_args to separate out the options for the
wrapper script itself from the rest that should simply be forwarded to
Ninja.

Tested: Ran pw-wrap-ninja with different values of --ui-update-rate and
  --ui-max-actions. Confirmed additional arguments were forwarded to
  Ninja.
Fixes: b/263979195
Change-Id: I14e39605999a36ba133f059626ddfcd4f454365b
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/125694
Commit-Queue: Eli Lipsitz <elipsitz@google.com>
Reviewed-by: Rob Mohr <mohrr@google.com>