pw_perf_test: Support overloaded functions in the simple macro
Switch to using a lambda for the PW_PERF_TEST_SIMPLE macro. That makes
it possible for the compiler to disambiguate between overloads of the
function under test, since it is called in the lambda.

Change-Id: Ifcc33708a1e3a7810ed95ef908991cbe7a94e3eb
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/125150
Commit-Queue: Auto-Submit <auto-submit@pigweed.google.com.iam.gserviceaccount.com>
Reviewed-by: Erik Gilling <konkers@google.com>
Pigweed-Auto-Submit: Wyatt Hepler <hepler@google.com>