pw_span: Remove pw_polyfill dependency
This dependency was added back when the type needed to support more
limited constexpr semantics for C++11 and/or C++14. Those language
versions are no longer targeted, and so the header is now just hanging
around as a ghost.

Note that compatibility_test.cc still depends on pw_polyfill. As a
result, I've added the relevant dependency for the module that builds
that test.

Also, pw_varint/varint.h was naughty and depended on pw_polyfill
transitively through pw_span, so I also have had to fix it.

Change-Id: I1f7817a331937b06e06e677e165977f002d62cfa
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/124873
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Anqi Dong <anqid@google.com>