pw_build: Avoid importing pty on Windows
The pw-wrap-ninja command relies on pty for intercepting and parsing
Ninja output. This doesn't work on Windows, so on Windows the script
simply runs Ninja without any wrapper.

As it turns out, the 'pty' module, while existing on Windows, breaks if
you attempt to actually import it. Since this module isn't used in the
script when it runs on Windows anyways, this commit guards its import
behind a platform check.

Change-Id: I73f58911f6e665b6a46d32a3acf600ee29760585
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/126941
Reviewed-by: Rob Mohr <mohrr@google.com>
Commit-Queue: Eli Lipsitz <elipsitz@google.com>