bazel: Build with both C++17 and C++20
Move the C++ standard version configuration from on-each-target to the
.bazelrc. The on-target settings have the highest precedence and cannot
be overridden. The .bazelrc settings have the _lowest_ precedence and
can be overridden, either on the command line or on individual targets.

Use this flexibility to build with both C++17 and C++20 in presubmit.

Bug: 268249509
Change-Id: I6399e4d95acbedc9549c2074eceb9c31c8118e00
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/128373
Commit-Queue: Ted Pudlik <tpudlik@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>
Reviewed-by: Fernando Garcia Bermudez <fgb@google.com>