pw_tokenizer: Update %p machinery
Changes how the `%p` machinery is handled to support more flags and
modifiers.

This version superficially looks the same as the old version, but allows
`+`, `-`, ` `, and user-specified width. The implementation is largely
the same as the machinery to workaround Python having a non-standard
octal `#` flag (implemented in the next CL).

Bug: b/265307572
Change-Id: Ib217a11d4c62bc5d079f5eb8938aca1fb8ee29e2
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/126836
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Greg Pataky <gregpataky@google.com>