pw_tokenizer: Add tests for %% specifier
This makes a decision to adhere to the C99 specification that `%` is a
specifier only when used exactly as `%%`, without any flag, width,
precision, or length modifiers. This was the current implementation
since Python 3 has issues with things like `%+%`, but it is worth noting
that both clang and gcc do allow any modifiers (and simply ignore them).

This can be changed in the future as long as documentation and tests are
updated.

Bug: b/265307572
Change-Id: I939d2b2cb7b551f564f3a8167906015ff557d208
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/126838
Commit-Queue: Greg Pataky <gregpataky@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>