pw_async: Create Task & FakeDispatcher facades
Create a facade for the Task class. The Task class now wraps a
NativeTask object that the backend implements.

Create Task backend for BasicDispatcher.

Create a facade for the FakeDispatcher class. The FakeDispatcher class
now wraps a NativeFakeDispatcher object that the backend implements.

Create a FakeDispatcher backend for BasicDispatcher.

Use a special value of 0 in NativeTask::interval_ instead of using an
optional to track whether a task is periodic. This reduces the size of
Task/NativeTask by 8 bytes.

Bug: b/254532947

Change-Id: I0ffdcedb55c241f356e87b86d603ada612a23cae
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/127029
Reviewed-by: Keir Mierle <keir@google.com>
Commit-Queue: Ben Lawson <benlawson@google.com>