pw_{doctor,env_setup}: Fix Rosetta setting parsing
This fixes two bugs in our use of Rosetta for Apple Silicon support:

1. `unames_are_equivalent` did not properly handle the case of running
on Apple Silicon *without* Rosetta. This led to the Pigweed environment
acting as if it was in an invalid state, even though the project was
designed to be in this state. This is now fixed.

2. Projects configure Rosetta support using `config.json` now, but
originally configured it by setting the `_PW_ROSETTA` env var in their
bootstrap script. After that transition, there remained one location
where we directly read the env var instead of reading `config.json`.
As a result, a pre-existing `_PW_ROSETTA` env var setting that didn't
match the one specified by a project's `config.json` could prevail
during env setup, leading to the wrong binaries being pulled down by
CIPD. An example of where this could occur: working in a project
where Rosetta was enabled, then navigating to a project where it
is disabled, without first deactivating or creating a new terminal
instance.

Change-Id: I59c37aee377bc669d8502c60040649b862d9f7e4
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/126461
Commit-Queue: Chad Norvell <chadnorvell@google.com>
Reviewed-by: Keir Mierle <keir@google.com>