pw_stream: Fix SocketStream compilation on macOS
SocketStream was using the MSG_NOSIGNAL flag with send to avoid getting
a SIGPIPE when the remote peer drops the connection. This works on Linux
only.

On macOS, you have to set a sockopt on the socket. This commit guards
the Linux-only code to only compile on Linux, and adds the equivalent
macOS code (guarded to only compile on macOS).

Change-Id: I731ebf7deea0ae2fcc043858b498badc1d399516
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/132010
Commit-Queue: Eli Lipsitz <elipsitz@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>