pw_stream: SocketServer: allow reusing the port
Previously if the server started listening on port 0 it would not use
SO_REUSEADDR, so if later on the server was restarted on the same
(resolved) port number it would fail to bind. In order to fix we are
moving setsocktopt() call so that it happens for port 0 as well.

Change-Id: Ic8119e364b405a0edab978d88e1c8d7497a5ca67
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/133991
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Oleg Shaldybin <olegsh@google.com>
Reviewed-by: Eli Lipsitz <elipsitz@google.com>