pw_i2c: make Device objects movable
Currently the Device object is not copyable or movable. This change
adds back the default move constructor. The Device is just a wrapper
around an address and Initiator reference, so it is safe to move.

Add documentation about the concurrency requirements of Initiator and
Device. Explain that Device is intended to represent ownership, and
that's why it is not copyable.

This change is intended to make it easier to construct objects which
depend on a Device. Previously the Device had to be allocated separately
and passed by reference to the constructor. This change allows using
temporary Device objects when constructing higher level objects.

Change-Id: I0671d4d16928f2657ac7269092e767171e5a6812
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/134610
Reviewed-by: Anthony DiGirolamo <tonymd@google.com>
Commit-Queue: Anton Markov <amarkov@google.com>