pw_sync: Make Borrowable<T> covariant in T
Allow converting Borrowable<U> to Borrowable<T> if U is a subclass of T.
This mirrors pointer and reference semantics w.r.t. base classes.

The motivation is to allow a Borrowable<U> or InlineBorrowable<U> to be
used where Borrowable<T> is accepted. This requires passing Borrowable
by value or const reference. Pass-by-value is preferred, because const
references can create confusion about what is const. Added documentation
to use pass-by-value.

Changed the acquire methods to be const, which is more consistent with
pointer semantics. This is required to use a Borrowable member value of
a const object.

Change-Id: I9aca5ac69e75ff1f6dfa7ea31a2dbf633f1259ad
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/134311
Reviewed-by: Ewout van Bekkum <ewout@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Anton Markov <amarkov@google.com>