pw_sync: Use Borrowable with value semantics
Change functions accepting Borrowable by reference to accept a value
instead. This change is always safe to make given the value semantics.

Don't change instance of returning Borrowable by reference, because
changing to return by value may break code that expects an lvalue.

Updated documentation for pw_spi to use a temporary Borrowable in the
example for creating a Device. This fits the spirit of the example.
Also fixed some typos in the example.

Change-Id: I624600a63aa5fe727631bdec486e81dcad080cf2
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/134312
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Anton Markov <amarkov@google.com>