pw_build: Better fallback error logging for ninja commands
Errors during the regen phase of a ninja invocation were not being
surfaced to the user in parallel builds. This adds a check to see if
any of the ninja errors are too short or only say 'FAILED:
build.ninja'. If true then the whole error log will be displayed on
the terminal as if it were a failure with any other command.

Change-Id: I3f846c8bf842c3b135bcd967fc16fce4c7bac243
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/135412
Reviewed-by: Erik Gilling <konkers@google.com>
Commit-Queue: Anthony DiGirolamo <tonymd@google.com>