pw_ide: Broaden build system support
In contrast to GN, which has a single top-level build that manages all
targets, CMake generates a separate Ninja build for each target. This is
actually the typical case, and GN's approach is atypical. pw_ide now
takes a more hands-off approach and simply finds compilation databases
regardless of how or where you generated them. It will use them as-is if
it can (the typical case), or process them if it needs to (the GN case).

This also changes the canonical compilation database from a symlink to a
a real file, which means we update the IDE's clangd settings when
changing targets instead. This allows clangd to automatically detect
when the database changes and restart itself, which it does not do for
symlinks. This means that in many cases, it will not be necessary to run
`pw ide sync` very frequently, since rebuilds that update the
compilation database will be automatically applied to clangd.

Bug: b/272566137
Bug: b/272566742
Change-Id: Id065fd2bfc0032702ce02339d9680ea50e9d722e
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/115445
Reviewed-by: Anthony DiGirolamo <tonymd@google.com>
Commit-Queue: Chad Norvell <chadnorvell@google.com>