pw_build: Update coverage_report.gni to support new -ffile-prefix-map
Shortly before the original coverage_report.gni CL landed, a CL went out
that updated the `-ffile-prefix-map` mappings. This broke the way we
were doing coverage report generation since we were relying on the old
file path mappings and the fact that `llvm-cov` would prepend the CWD to
the path before trying to resolve it (in this case, that was always the
build directory).

This CL updates to specify a `--compilation-dir` mapping to the root
project directory. To handle us mapping the build directory to `out`
always, we then use the `--path-equivalence` option to map that to the
current build directory so that the coverage report can correctly find
the files.

Change-Id: I191abea0bd1f34d2b1eeec8bc57f3f4eb86ffce2
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/138470
Reviewed-by: Gulfem Savrun Yeniceri <gulfem@google.com>
Reviewed-by: Carlos Chinchilla <cachinchilla@google.com>
Commit-Queue: Greg Pataky <gregpataky@google.com>
Presubmit-Verified: CQ Bot Account <pigweed-scoped@luci-project-accounts.iam.gserviceaccount.com>