pw_build: Fix BazelRule visibility
Prior to this CL, `BazelRule` had a `_visibility` attribute. This was
incorrect, as it was not set from the output of `bazel cquery`. This CL
removes that field, and its associated getter and setter used by tests.
Instead, visibility is treated in the same manner as other parsed
fields.

It also makes GnTarget smarter when adding visibility scopes, removing
scopes that are wholly within others.

Change-Id: Ib00affe252f8b2ee4a9f1142cbe648dff963c92f
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/143131
Reviewed-by: Carlos Chinchilla <cachinchilla@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>