pw_string: Only test strings of character types
This CL modifies the `TEST_FOR_TYPES` macro to remove integer types such
as `long long` and add character types such as `wchar_t`. The `char16_t`
and `char32_t` types are guarded by a check on C++11, and the `char8_t`
type is guarded by a check on C++20.

Non-character types are deprecated in LLVM as of 2022 and prevent
migration to LLVM 18:
https://discourse.llvm.org/t/deprecating-std-string-t-for-non-character-t/66779

Change-Id: I25f791b6a9011f9ea9fce51096f1db991420c805
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/146150
Reviewed-by: Prabhu Karthikeyan Rajasekaran <prabhukr@google.com>
Reviewed-by: Carlos Chinchilla <cachinchilla@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>