pw_rpc: Add destructor to stateful Call subclasses
b5f8cd9 moved waiting for callback completion from Call's destructors
into the destructors of some subclasses. This helped ensure that the
callbacks were completed prior to the destruction of fields referenced
by those callbacks.

However, that change missed a number of classes, and applied the
destructor to some classes where no extra state was carried (except
transitively through parent classes). This resulted in some bugs,
notably b/282138354.

This change moves the DestroyClientCall and DestroyServerCall
invocations into two places:
 - The destructors of any stateful Call subclasses (those subclasses
   whose fields may be referenced by the callbacks).
 - The destructors of ServerCall and ClientCall (to guard the stateful
   `Call` class appropriately in the case of classes whose heirarchy
   would not otherwise have invoked `DestroyXXXCall`).

This results in some duplication, notably between classes where multiple
layers in the inheritance tree are stateful: each stateful layer will
`DestroyXXXCall`. However, the alternative is leaving intermediate
classes unprotected, which makes it undefined behavior to create an
instance of such an intermediate class, and makes it hard to track
whether the final subclasses are properly protected: they must include a
destructor if and only if their parents had state and did *not* include
a destructor. This condition is much too complex to validate, especially
given that Call has 40+ subclasses, many with four-deep inheritance.

After this CL, it should be easy to check a class for correctness:
it must include a `DestroyXXXCall`-invoking destructor if it has a field
whose state may be referenced from a callback.

Bug: b/282138354
Change-Id: I2fb43b254a653873de102cedf7becfcc8cfd8c77
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/148110
Commit-Queue: Taylor Cramer <cramertj@google.com>
Reviewed-by: Aaron Green <aarongreen@google.com>