pw_log_zephyr: Add EOT character to help with detokenizing on Zephyr
Zephyr console may output raw text along with the tokenized messaging.
This causes issues in determining when the end of a base64 message
occurs to process a token. Adding an EOT to the end of a base64 message
helps encapsulate the entire token or when we're receiving a token.
Upon receiving the EOT the token can be fully decoded without errors.

Change-Id: If74c64abcffd591f6c88968155c7954f996c1bcd
Signed-off-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/153751
Reviewed-by: Taylor Cramer <cramertj@google.com>
Commit-Queue: Auto-Submit <auto-submit@pigweed.google.com.iam.gserviceaccount.com>