pw_protobuf_compiler: Transition to our own proto compiler rules
In order to do this, we need to add support for protos that depend on
other protos. This adds the dependent proto paths to protoc invocation.

However, this has a side-effect in where the output files get put. They
are now always placed relative to the root, instead of next to their
source .proto files. Handle this by declaring the output files based on
the targets output path, and adjust include paths accordingly.

Also add support for options files in external repositories and enable
some tests that use .options files.

Bug: b/234874064
Bug: b/234873954
Change-Id: I49acf98396652d3ba7a5b547b6f868a48def0238
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/157033
Commit-Queue: Austin Foxley <afoxley@google.com>
Reviewed-by: Ted Pudlik <tpudlik@google.com>