pw_build: Add python_action_test
This CL adds a GN template for tests that use a Python script. This
template wraps `pw_python_action` and adds test metadata.

Action tests are distinguished from integration tests in that the
former run quickly enough to be part of the default set of tests. A
subsequent CL will use this template to add metadata for integration
tests as well.

Change-Id: I7706eaac51596047a70a590d2216799d8496b444
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/154552
Reviewed-by: Armando Montanez <amontanez@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>