pw_hdlc: Mitigate errors on Python background thread
Prior to this change, when an exception was encountered by the
DataReaderAndExecutor background thread, it would hammer on _read()
indefinitely until the program was shut down.

With the addition of logging in pwrev/155814, this caused error messages
to be spammed to the log very rapidly.

This commit mitigates the issue, without significantly changing the API,
by exiting the reader thread when an exception is encountered.

This also adds a stop() method and close() methods on HdlcRpcClient
to signal the thread to stop. Currently, this is opportunistic as it is
unable to interrupt a blocking read() call.

pw_hdlc.rpc:
- Move reader thread into DataReaderAndExecutor.
- Make RpcClient classes a context manager.
- Improve logging of read errors to include backtrace.
- Add a test for these shutdown cases.

Test: ninja -C out
Bug: 293595266
Change-Id: I7b6fe9607e47cd72558b58de9c93f13c97b7582f
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/162712
Reviewed-by: Carlos Chinchilla <cachinchilla@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>
Pigweed-Auto-Submit: Jonathon Reinhart <jrreinhart@google.com>
Commit-Queue: Auto-Submit <auto-submit@pigweed-service-accounts.iam.gserviceaccount.com>
Presubmit-Verified: CQ Bot Account <pigweed-scoped@luci-project-accounts.iam.gserviceaccount.com>