pw_function: Add Allocator injection
The Fuchsia change
(https://fuchsia-review.googlesource.com/c/fuchsia/+/909692) requires
Pigweed's copies of the 2 Fuchsia function.h files to be updated to the
new code in order to pass presubmits, due to a circular dependency, so
manually modify them in this CL, and then run the copybara script in a
followup CL once the Fuchsia change is merged.

Test: The Fuchsia change has been tested there and the new test source
will be updated in Pigweed with the copybara run.

Change-Id: I6696e90c7c4b595dfdc1f9fc2ee23d844d51afff
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/170190
Commit-Queue: Leonard Zgrablic <lzgrablic@google.com>
Reviewed-by: Armando Montanez <amontanez@google.com>