pw_chre: Add barebones CHRE
NOTE: This is a very work-in-progress implementation.

Add GN rules for the CHRE codebase into //third_party/CHRE.
This also adds a pw_chre_PLATFORM_BACKEND gn argument for the
platform backend for CHRE.

Add a pw_chre directory that includes all of the class and
function definitions required to compile CHRE. Some of this
connects to pigweed functionality, some of it is stubbed out.

//pw_chre:chre_simulator uses pw_system to run CHRE on
pigweed on the host, and it compiles in a number of CHRE's
example nanoapps. There is enough functionality that these
nanoapps are able to run.

Bug: 294106526

Change-Id: I73ab1a83ccefbe4a7d0f9ba7a5421c3f3391a937
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/162510
Reviewed-by: Kayce Basques <kayce@google.com>
Reviewed-by: Armando Montanez <amontanez@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Aaron Green <aarongreen@google.com>