pw_libc: Fix printf for newer llvm-libc commits
This file was removed from the upstream implementation.

There's no need to check if the revision is after when the file
was removed because we have no users of pw_libc yet. In the future
these kinds of transitions will be more difficult, but for now we
can assume any user of pw_libc is using a revision of llvm-libc
after the removal happened.

Change-Id: I1797ad277586d06ded175b3a7014e024abb88bfb
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/170831
Commit-Queue: Alex Brachet <abrachet@google.com>
Reviewed-by: Armando Montanez <amontanez@google.com>