docs: Start the glossary
Creates a new glossary (//docs/glossary.rst) and starts it off with a
definition for "module". Updates other docs to link to the glossary
definition.

Over the long run we may want a more structured approach. Maybe we
store the glossary definitions in a different format (e.g. JSON) to
make it easier to auto-link to the glossary in the rest of the docs.
For now, we should just start with the simplest solution and only
add complexity when necessary.

Change-Id: I060278abc2ff0661f381e50cafcc5e3a0b73a191
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/172952
Reviewed-by: Chad Norvell <chadnorvell@google.com>
Reviewed-by: Kayce Basques <kayce@google.com>
Commit-Queue: Kayce Basques <kayce@google.com>