pw_perf_test: Gate on pw_chrono_SYSTEM_TIMER_BACKEND
A backend for the system timer is required for the chrono timer test.
This changes the test to gate on that rather than the system clock, as a
target with a system clock set up might not necessarily have a timer
backend set up yet.

Change-Id: Ifc01c65228e48e4a5c15878c550e475d5c48b708
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/174650
Commit-Queue: Auto-Submit <auto-submit@pigweed-service-accounts.iam.gserviceaccount.com>
Pigweed-Auto-Submit: Armando Montanez <amontanez@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>