pw_presubmit: Handle missing upstream better
If we try to find the commit at the remote tracking branch but are
unsuccessful, use 'HEAD~10'. This is excessive for individual CLs, but
not overly so, and it should be enough for most stacks of CLs.

Bug: b/282808936
Change-Id: Iba2b0c2d4cc571fa6c0f9aca6227b14bcf976276
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/177038
Pigweed-Auto-Submit: Rob Mohr <mohrr@google.com>
Reviewed-by: Taylor Cramer <cramertj@google.com>
Presubmit-Verified: CQ Bot Account <pigweed-scoped@luci-project-accounts.iam.gserviceaccount.com>
Commit-Queue: Auto-Submit <auto-submit@pigweed-service-accounts.iam.gserviceaccount.com>