pw_emu: Add support for substitutions in config entries
For now only $pw_env{} is supported. Note that the pw_env_setup also
performs this substitution, but this does not work when using
pw_emu configuration fragments.

Ideally this should be done in a generic fashion for all Pigweed
modules (b/309161337) but this is out of scope for this CL which aims
at allowing downstream projects to perform substitution for emulation
target fragments (e.g. for using a project specific qemu binary
relative to the project root).

This CL also adds unit tests for pw_emu.core.Config.

Bug: 308793747
Change-Id: I7f9208276b3981754316d674e0b25721c9beb165
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/179150
Commit-Queue: Octavian Purdila <tavip@google.com>
Reviewed-by: Jonathon Reinhart <jrreinhart@google.com>