pw_system: Add tracing to the demo system
Adds a sample implementation of the trace service with built on top of
the flat filesystem and transfer services. The trace data is backed by a
persistent buffer, again as an sample implementation.

Also fixes an issue with pw_transfer failing to compile on systems with
different periods.

Note:
- the bazel build doesn't link due to tracing not being supported in the
  bazel build(b/260641850), so tracing is disabled in the sample bazel build.
- the cmake build doesn't compile, but it was updated we new targets so
as not to dig the hole any deeper.

Change-Id: I483d049d486647c10be3db8f104ea2419a31a312
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/168834
Commit-Queue: Dave Roth <davidroth@google.com>
Presubmit-Verified: CQ Bot Account <pigweed-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Armando Montanez <amontanez@google.com>
Reviewed-by: Alexei Frolov <frolv@google.com>